Backprop.Learn

Documentation

data Chain :: [Type] -> [Type] -> Type -> Type -> Type where

Chain components linearly, retaining the ability to deconstruct at a later time.

Constructors

CNil :: Chain '[] '[] a a

(:~>) :: (Learn p a b l, Num b) => l -> Chain ls ps b c -> Chain (l ': ls) (p ': ps) a c infixr 5

(~:++) :: forall ls ks ps qs a b c. (Known Length ps, Known Length qs, ListC (Num <$> ps), ListC (Num <$> qs)) => Chain ls ps a b -> Chain ks qs b c -> Chain (ls ++ ks) (ps ++ qs) a c infixl 5

data LearnFunc :: Type -> Type -> Type -> Type where

Data type representing trainable models.

Useful for performant composition, but you lose the ability to decompose parts.

Constructors

LF :: {..} -> LearnFunc p a b

Fields

_lfInitParam :: forall m. PrimMonad m => Gen (PrimState m) -> m p

_lfRunFixed :: forall s. Reifies s W => BVar s p -> BVar s a -> BVar s b

_lfRunStoch :: forall m s. (PrimMonad m, Reifies s W) => Gen (PrimState m) -> BVar s p -> BVar s a -> m (BVar s b)

learnFunc :: Learn p a b l => l -> LearnFunc p a b

(~>) :: (Num p, ListC (Num <$> ps), Known Length ps) => LearnFunc p a b -> LearnFunc (T ps) b c -> LearnFunc (T (p ': ps)) a c infixr 5

nilLF :: LearnFunc (T '[]) a a

(~!++) :: LearnFunc NoParam a b -> LearnFunc (T ps) b c -> LearnFunc (T ps) a c infixr 5

(~++!) :: LearnFunc (T ps) a b -> LearnFunc NoParam b c -> LearnFunc (T ps) a c infixr 5

(~++) :: forall ps qs a b c. (Known Length ps, Known Length qs, ListC (Num <$> ps), ListC (Num <$> qs)) => LearnFunc (T ps) a b -> LearnFunc (T qs) b c -> LearnFunc (T (ps ++ qs)) a c

data l :.~ k infixr 7

Simple composition of Learn instances