Backprop.Learn.Class

Documentation

class Learn a b l | l -> a, l -> b where

Class for models that can be trained using gradient descent

An instance l of Learn a b is parameterized by p, takes a as input, and returns b as outputs. l can be thought of as a value containing the hyperparmaeters of the model.

Minimal complete definition

runLearn

Associated Types

type LParamMaybe l :: Maybe Type

The trainable parameters of model l.

By default, is 'Nothing. To give a type for learned parameters p, use the type 'Just p

type LStateMaybe l :: Maybe Type

The type of the state of model l. Used for things like recurrent neural networks.

By default, is 'Nothing. To give a type for state s, use the type 'Just s.

Most models will not use state, training algorithms will only work if LStateMaybe is 'Nothing. However, models that use state can be converted to models that do not using Unroll; this can be done before training.

Methods

initParam :: PrimMonad m => l -> Gen (PrimState m) -> LParam m l

Initialize parameters, given the hyperparameters in l.

initParam :: LParamMaybe l ~ Nothing => l -> Gen (PrimState m) -> LParam m l

Initialize parameters, given the hyperparameters in l.

initState :: PrimMonad m => l -> Gen (PrimState m) -> LState m l

Initialize state, given the hyperparameters in l.

initState :: LStateMaybe l ~ Nothing => l -> Gen (PrimState m) -> LState m l

Initialize state, given the hyperparameters in l.

runLearn :: Reifies s W => l -> LParam (BVar s) l -> BVar s a -> LState (BVar s) l -> (BVar s b, LState (BVar s) l)

Run the model itself, deterministically.

If your model has no state, you can define this conveniently using stateless.

runLearnStoch :: (Reifies s W, PrimMonad m) => l -> Gen (PrimState m) -> LParam (BVar s) l -> BVar s a -> LState (BVar s) l -> m (BVar s b, LState (BVar s) l)

Run a model in stochastic mode.

If model is inherently non-stochastic, a default implementation is given in terms of runLearn.

If your model has no state, you can define this conveniently using statelessStoch.

type LParamOf l = FromJust (ShowType l :<>: Text " has no trainable parameters") (LParamMaybe l)

The trainable parameter type of a model. Will be a compile-time error if the model has no trainable parameters.

type LStateOf l = FromJust (ShowType l :<>: Text " has no trainable parameters") (LStateMaybe l)

The state type of a model. Will be a compile-time error if the model has no state.

type NoParam l = LParamMaybe l ~ Nothing

Constraint specifying that a given model has no trainabale parameters.

type NoState l = LStateMaybe l ~ Nothing

Constraint specifying that a given model has no state.

type LParam f l = Mayb f (LParamMaybe l)

Is N_ if there is l has no trainable parameters; otherwise is J_ with f p, for trainable parameter type p.

type LState f l = Mayb f (LStateMaybe l)

Is N_ if there is l has no state; otherwise is J_ with f s, for state type s.

stateless :: (a -> b) -> a -> s -> (b, s)

Useful for defining runLearn if your model has no state.

statelessM :: Functor m => (a -> m b) -> a -> s -> m (b, s)

Useful for defining runLearnStoch if your model has no state.

runLearnStateless :: (Learn a b l, Reifies s W, NoState l) => l -> LParam (BVar s) l -> BVar s a -> BVar s b

runLearnStochStateless :: (Learn a b l, Reifies s W, NoState l, PrimMonad m) => l -> Gen (PrimState m) -> LParam (BVar s) l -> BVar s a -> m (BVar s b)

data Mayb :: (k -> Type) -> Maybe k -> Type where

Constructors

N_ :: Mayb f Nothing

J_ :: f a -> Mayb f (Just a)

Bundled Patterns

pattern J_I :: a -> Mayb I (Just a)

fromJ_ :: Mayb f (Just a) -> f a

type family MaybeC (c :: k -> Constraint) (m :: Maybe k) :: Constraint where ...

Equations

MaybeC c (Just a) = c a

MaybeC c Nothing = ()

type KnownMayb = Known (Mayb P)

knownMayb :: KnownMayb p => Mayb P p