Backprop.Learn.Component.Combinator

Documentation

data Chain :: [Type] -> [Type] -> [Type] -> Type -> Type -> Type where

Chain components linearly, retaining the ability to deconstruct at a later time.

Constructors

CNil :: Chain '[] '[] '[] a a

(:~>) :: (Learn a b l, KnownMayb (LParamMaybe l), KnownMayb (LStateMaybe l)) => l -> Chain ls ps ss b c -> Chain (l ': ls) (MaybeToList (LParamMaybe l) ++ ps) (MaybeToList (LStateMaybe l) ++ ss) a c infixr 5

(~++) :: forall ls ms ps qs ss ts a b c. Chain ls ps ss a b -> Chain ms qs ts b c -> Chain (ls ++ ms) (ps ++ qs) (ss ++ ts) a c

Appending Chain

chainParamLength :: Chain ls ps ss a b -> Length ps

chainStateLength :: Chain ls ps ss a b -> Length ss

data LearnFunc :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

Data type representing trainable models.

Useful for performant composition, but you lose the ability to decompose parts.

Constructors

LF :: {..} -> LearnFunc p s a b

Fields

_lfInitParam :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m p

_lfInitState :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m s

_lfRunLearn :: forall q. Reifies q W => Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> (BVar q b, Mayb (BVar q) s)

_lfRunLearnStoch :: forall m q. (PrimMonad m, Reifies q W) => Gen (PrimState m) -> Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> m (BVar q b, Mayb (BVar q) s)

learnFunc :: Learn a b l => l -> LearnFunc (LParamMaybe l) (LStateMaybe l) a b

(.~) :: forall ps qs ss ts a b c. (ListC (Num <$> ps), ListC (Num <$> qs), ListC (Num <$> ss), ListC (Num <$> ts), ListC (Num <$> (ss ++ ts)), Known Length ps, Known Length qs, Known Length ss, Known Length ts) => LearnFunc (Just (T ps)) (Just (T ss)) b c -> LearnFunc (Just (T qs)) (Just (T ts)) a b -> LearnFunc (Just (T (ps ++ qs))) (Just (T (ss ++ ts))) a c

Compose two LearnFunc on lists.

nilLF :: LearnFunc (Just (T '[])) (Just (T '[])) a a

Identity of .~

onlyLF :: forall p s a b. (KnownMayb p, MaybeC Num p, KnownMayb s, MaybeC Num s) => LearnFunc p s a b -> LearnFunc (Just (T (MaybeToList p))) (Just (T (MaybeToList s))) a b

LearnFunc with singleton lists; meant to be used with .~