Backprop.Learn.Component.Function

Documentation

data ParamFunc p a b

A ParamFunc p a b is a parameterized function from a to b, potentially with trainable parameter p.

Constructors

PF

Fields

_pfInit :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m p

_pfFunc :: forall s. Reifies s W => Mayb (BVar s) p -> BVar s a -> BVar s b

Bundled Patterns

pattern FF :: (forall s. Reifies s W => BVar s a -> BVar s b) -> FixedFunc a b

type FixedFunc = ParamFunc Nothing

An unparameterized function. Has a Category instance.

runFixedFunc :: FixedFunc a b -> forall s. Reifies Type s W => BVar s a -> BVar s b

learnParam :: forall l a b. (Learn a b l, NoState l) => l -> ParamFunc (LParamMaybe l) a b

Create a ParamFunc from any instance of Learn that does not have state.

dimapPF :: FixedFunc a b -> FixedFunc c d -> ParamFunc p b c -> ParamFunc p a d

Pre- and post-compose a ParamFunc

lmapPF :: FixedFunc a b -> ParamFunc p b c -> ParamFunc p a c

Precompose a ParamFunc

rmapPF :: FixedFunc b c -> ParamFunc p a b -> ParamFunc p a c

Postcompose a ParamFunc

compPF :: (Num p, Num q) => ParamFunc (Just p) a b -> ParamFunc (Just q) b c -> ParamFunc (Just (T2 p q)) a c

Compose two ParamFuncs sequentially

parPF :: (Num p, Num q, Num a, Num b, Num c, Num d) => ParamFunc (Just p) a c -> ParamFunc (Just q) b d -> ParamFunc (Just (T2 p q)) (T2 a b) (T2 c d)

Compose two ParamFuncs in parallel

softMax :: KnownNat i => FixedFunc (R i) (R i)

Softmax normalizer

logisticFunc :: Floating a => FixedFunc a a

Logistic function, typically used as an activation function

scaleFunc :: Num a => a -> FixedFunc a a

Scaling function

tanhFunc :: Floating a => FixedFunc a a

Tanh function

mapFunc :: KnownNat i => (forall s. Reifies s W => BVar s Double -> BVar s Double) -> FixedFunc (R i) (R i)

Map a differentiable function over every item in an R

reLU :: KnownNat i => FixedFunc (R i) (R i)

Rectified linear unit activation function

eLU :: KnownNat i => FixedFunc (R i) (R i)

Exponential linear unit activation function

pScale :: (KnownNat i, ContGen d) => d -> ParamFunc (Just Double) (R i) (R i)

Scaling function, but with a trainable scaling parameter.

pMap :: KnownNat i => (forall m. PrimMonad m => Gen (PrimState m) -> m p) -> (forall s. Reifies s W => BVar s p -> BVar s Double -> BVar s Double) -> ParamFunc (Just p) (R i) (R i)

Map a parameterized differentiable function over ever item in an R. The parameteri is trainable.

(.-) :: forall ps qs a b c. (ListC (Num <$> ps), ListC (Num <$> qs), Known Length ps, Known Length qs) => ParamFunc (Just (T ps)) b c -> ParamFunc (Just (T qs)) a b -> ParamFunc (Just (T (ps ++ qs))) a c infixr 9

Compose two ParamFuncs on lists.

nilPF :: ParamFunc (Just (T '[])) a a

The identity of .-

onlyPF :: forall p a b. (KnownMayb p, MaybeC Num p) => ParamFunc p a b -> ParamFunc (Just (T (MaybeToList p))) a b

ParamFunc taking a singleton list; meant to be used with .-