Backprop.Learn.Component.Parameter

Documentation

data DeParam :: Type -> Type -> Type where

Convert a model with trainabile parameters into a model without any trainable parameters.

The parameters are instead fixed (or stochastic, with a fixed distribution), and the model is treated as an untrainable function.

Constructors

DP :: {..} -> DeParam p l

Fields

_dpParam :: p

_dpParamStoch :: forall m. PrimMonad m => Gen (PrimState m) -> m p

_dpLearn :: l

dpDeterm :: p -> l -> DeParam p l

Create a DeParam from a deterministic, non-stochastic parameter.

data ReParam :: Type -> Maybe Type -> Type -> Type

Pre-apply a function to the parameter before the original model sees it. A ReParam p q turns a model taking p into a model taking q.

Note that a ReParam p 'Nothing is essentially the same as DeParam p; however, defining one in terms of the other is a bit inconvenient, so they are provided as separate types.

rpDeterm :: (forall m. PrimMonad m => Gen (PrimState m) -> Mayb m q) -> (forall s. Reifies s W => Mayb (BVar s) q -> BVar s p) -> l -> ReParam p q l

Create a ReParam from a deterministic, non-stochastic transformation function.