Backprop.Learn.Component.State

Documentation

newtype DeState :: Type -> Type where

Make a model stateless by converting the state to a trained parameter, and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See FixState, as well.

Its parameters are:

If l has no parameter or state, nothing.

If l has parameters but no state, just the original parameters.

If l has no parameters but state, just the initial state.

If l has parameters and state, a T2 of the parameter and initial state.

However, this is really only meant to be used with types that have state. If used with a type with no state, this is essentially a no-op/identity.

Constructors

DeState :: {..} -> DeState l

Fields

getDestate :: l

data FixState :: Maybe Type -> Type -> Type where

Make a model stateless by pre-applying a fixed state and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See DeState, as well.

If used with a type with no state, is essentially a no-op/identity.

Constructors

FS :: {..} -> FixState s l

Fields

_fsLearn :: l

_fsInitState :: Mayb I s

newtype Unroll :: Nat -> Type -> Type where

Unroll a stateful model into one taking a vector of sequential inputs.

Useful when used before DeState or FixState. See UnrollDeState and UnrollFixState.

Constructors

Unroll :: {..} -> Unroll n l

Fields

getUnroll :: l

type UnrollDeState n l = DeState (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and treat the initial state as a trained parameter.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollDeState n l)'

    type LParamMaybe (UnrollFixState n l) = 'TupMaybe (LParamMaybe l) (LStateMaybe l)
    type LStateMaybe (UnrollFixState n l) = 'Nothing

type UnrollFixState n l = FixState (LParamMaybe l) (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and fix the initial state.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollFixState n l)'

    type LParamMaybe (UnrollFixState n l) = LParamMaybe l
    type LStateMaybe (UnrollFixState n l) = 'Nothing