Backprop.Learn.Initialize

Documentation

class Initialize p where

Class for types that are basically a bunch of Doubles, which can be initialized with a given identical and independent distribution.

Minimal complete definition

Nothing

Methods

initialize :: (ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

initialize :: (ADTRecord p, Constraints p Initialize, ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

gInitialize :: (ADTRecord p, Constraints p Initialize, ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

initialize for any instance of Generic.

initializeNormal

Arguments

:: (Initialize p, PrimMonad m)

=> Double
standard deviation

-> Gen (PrimState m)

-> m p

Helper over inititialize for a gaussian distribution centered around zero.

initializeSingle :: (ContGen d, PrimMonad m, Fractional p) => d -> Gen (PrimState m) -> m p

initialize definition if p is a single number.

Reshape

reshapeR :: forall i j d m. (ContGen d, PrimMonad m, KnownNat i, KnownNat j) => d -> Gen (PrimState m) -> R i -> m (R j)

Reshape a vector to have a different amount of items If the matrix is grown, new weights are initialized according to the given distribution.

reshapeLRows :: forall i j n d m. (ContGen d, PrimMonad m, KnownNat n, KnownNat i, KnownNat j) => d -> Gen (PrimState m) -> L i n -> m (L j n)

Reshape a matrix to have a different amount of rows If the matrix is grown, new weights are initialized according to the given distribution.

reshapeLCols :: forall i j n d m. (ContGen d, PrimMonad m, KnownNat n, KnownNat i, KnownNat j) => d -> Gen (PrimState m) -> L n i -> m (L n j)

Reshape a matrix to have a different amount of columns. If the matrix is grown, new weights are initialized according to the given distribution.