Backprop.Learn.Loss

Loss functions

type Loss a = forall s. Reifies s W => a -> BVar s a -> BVar s Double

crossEntropy :: KnownNat n => Loss (R n)

squaredError :: Loss Double

absError :: Loss Double

totalSquaredError :: (Num (t Double), Foldable t, Functor t) => Loss (t Double)

squaredErrorV :: KnownNat n => Loss (R n)

totalCov :: (Num (t Double), Foldable t, Functor t) => Loss (t Double)

Sum of covariances between matching components. Not sure if anyone uses this.

Manipulate loss functions

scaleLoss :: Double -> Loss a -> Loss a

Scale the result of a loss function.

sumLoss :: (Traversable t, Applicative t, Num a) => Loss a -> Loss (t a)

sumLossDecay :: forall n a. (KnownNat n, Num a) => Double -> Loss a -> Loss (Vector n a)

lastLoss :: (KnownNat (n + 1), Num a) => Loss a -> Loss (Vector (n + 1) a)

zipLoss :: (Traversable t, Applicative t, Num a) => t Double -> Loss a -> Loss (t a)

t2Loss

Arguments

:: (Num a, Num b)

=> Loss a
loss on first component

-> Loss b
loss on second component

-> Loss (T2 a b)

Lift and sum a loss function over the components of a T2.

t3Loss

Arguments

:: (Num a, Num b, Num c)

=> Loss a
loss on first component

-> Loss b
loss on second component

-> Loss c
loss on third component

-> Loss (T3 a b c)

Lift and sum a loss function over the components of a T3.

Regularization

type Regularizer p = forall s. Reifies s W => BVar s p -> BVar s Double

A regularizer on parameters

l2Reg

Arguments

:: (Metric Double p, Num p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

L2 regularization

\[ \sum_w \frac{1}{2} w^2 \]

l1Reg

Arguments

:: (Metric Double p, Num p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

L1 regularization

\[ \sum_w \lvert w \rvert \]

noReg :: Regularizer p

No regularization

Manipulate regularizers

addReg :: Regularizer p -> Regularizer p -> Regularizer p

Add together two regularizers

scaleReg :: Double -> Regularizer p -> Regularizer p

Scale a regularizer's influence