Backprop.Learn.Model.Class

Documentation

class Learn a b l | l -> a b where

Class for models that can be trained using gradient descent

An instance l of Learn a b is parameterized by p, takes a as input, and returns b as outputs. l can be thought of as a value containing the hyperparmaeters of the model.

Minimal complete definition

runLearn

Associated Types

type LParamMaybe l :: Maybe Type

The trainable parameters of model l.

By default, is 'Nothing. To give a type for learned parameters p, use the type 'Just p

type LStateMaybe l :: Maybe Type

The type of the state of model l. Used for things like recurrent neural networks.

By default, is 'Nothing. To give a type for state s, use the type 'Just s.

Most models will not use state, training algorithms will only work if LStateMaybe is 'Nothing. However, models that use state can be converted to models that do not using Unroll; this can be done before training.

Methods

initParam :: PrimMonad m => l -> Gen (PrimState m) -> LParam_ m l

Initialize parameters, given the hyperparameters in l.

Default definition provided for models with no state.

initParam :: NoParam l => l -> Gen (PrimState m) -> LParam_ m l

Initialize parameters, given the hyperparameters in l.

Default definition provided for models with no state.

initState :: PrimMonad m => l -> Gen (PrimState m) -> LState_ m l

Initialize state, given the hyperparameters in l.

Default definition provided for models with no state.

initState :: NoState l => l -> Gen (PrimState m) -> LState_ m l

Initialize state, given the hyperparameters in l.

Default definition provided for models with no state.

runLearn :: Reifies s W => l -> LParam_ (BVar s) l -> BVar s a -> LState_ (BVar s) l -> (BVar s b, LState_ (BVar s) l)

Run the model itself, deterministically.

If your model has no state, you can define this conveniently using stateless.

runLearnStoch :: (Reifies s W, PrimMonad m) => l -> Gen (PrimState m) -> LParam_ (BVar s) l -> BVar s a -> LState_ (BVar s) l -> m (BVar s b, LState_ (BVar s) l)

Run a model in stochastic mode.

If model is inherently non-stochastic, a default implementation is given in terms of runLearn.

If your model has no state, you can define this conveniently using statelessStoch.

type LParam l = FromJust (ShowType l :<>: Text " has no trainable parameters") (LParamMaybe l)

The trainable parameter type of a model. Will be a compile-time error if the model has no trainable parameters.

type LState l = FromJust (ShowType l :<>: Text " has no trainable parameters") (LStateMaybe l)

The state type of a model. Will be a compile-time error if the model has no state.

type family LParams (ls :: [Type]) :: [Type] where ...

List of parameters of Learn instances

Equations

LParams '[] = '[]

LParams (l ': ls) = MaybeToList (LParamMaybe l) ++ LParams ls

type family LStates (ls :: [Type]) :: [Type] where ...

List of states of Learn instances

Equations

LStates '[] = '[]

LStates (l ': ls) = MaybeToList (LStateMaybe l) ++ LStates ls

type NoParam l = LParamMaybe l ~ Nothing

Constraint specifying that a given model has no trainabale parameters.

type NoState l = LStateMaybe l ~ Nothing

Constraint specifying that a given model has no state.

type LParam_ f l = Mayb f (LParamMaybe l)

Is N_ if there is l has no trainable parameters; otherwise is J_ with f p, for trainable parameter type p.

type LState_ f l = Mayb f (LStateMaybe l)

Is N_ if there is l has no state; otherwise is J_ with f s, for state type s.

stateless :: (a -> b) -> a -> s -> (b, s)

Useful for defining runLearn if your model has no state.

statelessM :: Functor m => (a -> m b) -> a -> s -> m (b, s)

Useful for defining runLearnStoch if your model has no state.

runLearnStateless :: (Learn a b l, Reifies s W, NoState l) => l -> LParam_ (BVar s) l -> BVar s a -> BVar s b

runLearnStochStateless :: (Learn a b l, Reifies s W, NoState l, PrimMonad m) => l -> Gen (PrimState m) -> LParam_ (BVar s) l -> BVar s a -> m (BVar s b)

data Mayb :: (k -> Type) -> Maybe k -> Type where

Constructors

N_ :: Mayb f Nothing

J_ :: !(f a) -> Mayb f (Just a)

Bundled Patterns

pattern J_I :: a -> Mayb I (Just a)

fromJ_ :: Mayb f (Just a) -> f a

type family MaybeC (c :: k -> Constraint) (m :: Maybe k) :: Constraint where ...

Equations

MaybeC c (Just a) = c a

MaybeC c Nothing = ()

type KnownMayb = Known (Mayb P)

knownMayb :: KnownMayb p => Mayb P p

newtype I a

Constructors

I

Fields

getI :: a

data SomeLearn :: Type -> Type -> Type where

Existential wrapper for learnable model, representing a trainable function from a to b.

Constructors

SL :: (Learn a b l, Typeable l, KnownMayb (LParamMaybe l), KnownMayb (LStateMaybe l), MaybeC Num (LParamMaybe l), MaybeC Num (LStateMaybe l), MaybeC (Metric Double) (LParamMaybe l), MaybeC (Metric Double) (LStateMaybe l), MaybeC NFData (LParamMaybe l), MaybeC NFData (LStateMaybe l)) => l -> SomeLearn a b