Backprop.Learn.Model.Combinator

Documentation

data Chain :: [Type] -> Type -> Type -> Type where

Chain components linearly, retaining the ability to deconstruct at a later time.

Constructors

CNil :: Chain '[] a a

(:~>) :: (Learn a b l, KnownMayb (LParamMaybe l), KnownMayb (LStateMaybe l)) => l -> Chain ls b c -> Chain (l ': ls) a c infixr 5

(~++) :: forall ls ms a b c. Chain ls a b -> Chain ms b c -> Chain (ls ++ ms) a c

Appending Chain

chainParamLength :: Chain ls a b -> Length (LParams ls)

chainStateLength :: Chain ls a b -> Length (LStates ls)

data LearnFunc :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

Data type representing trainable models.

Useful for performant composition, but you lose the ability to decompose parts.

Constructors

LF :: {..} -> LearnFunc p s a b

Fields

_lfInitParam :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m p

_lfInitState :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m s

_lfRunLearn :: forall q. Reifies q W => Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> (BVar q b, Mayb (BVar q) s)

_lfRunLearnStoch :: forall m q. (PrimMonad m, Reifies q W) => Gen (PrimState m) -> Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> m (BVar q b, Mayb (BVar q) s)

learnFunc :: Learn a b l => l -> LearnFunc (LParamMaybe l) (LStateMaybe l) a b

data LMap :: Type -> Type -> Type -> Type where

Pre-compose a pure parameterless function to a model.

An LMap b a takes a model from b and turns it into a model from a.

Constructors

LM :: {..} -> LMap b a l

Fields

_lmPre :: forall s. Reifies s W => BVar s a -> BVar s b

_lmLearn :: l

data RMap :: Type -> Type -> Type -> Type where

Post-compose a pure parameterless function to a model.

An Rmap b c takes a model returning b and turns it into a model returning c.

Constructors

RM :: {..} -> RMap b c l

Fields

_rmPost :: forall s. Reifies s W => BVar s b -> BVar s c

_rmLearn :: l

type Dimap b c a d l = RMap c d (LMap b a l)

Pre- and post-compose pure parameterless functions to a model.

A Dimap b c a d takes a model from b to c and turns it into a model from a to d.

instance Learn b c => Learn a d (Dimap b c a d l) where
    type LParamMaybe (Dimap b c a d l) = LParamMaybe l
    type LStateMaybe (Dimap b c a d l) = LStateMaybe l

pattern DM :: (forall s. Reifies s W => BVar s a -> BVar s b) -> (forall s. Reifies s W => BVar s c -> BVar s d) -> l -> Dimap b c a d l

Constructor for Dimap

_dmPre :: Dimap b c a d l -> forall s. Reifies s W => BVar s a -> BVar s b

_dmPost :: Dimap b c a d l -> forall s. Reifies s W => BVar s c -> BVar s d

_dmLearn :: Dimap b c a d l -> l

(.~) :: forall ps qs ss ts a b c. (ListC (Num <$> ps), ListC (Num <$> qs), ListC (Num <$> ss), ListC (Num <$> ts), ListC (Num <$> (ss ++ ts)), Known Length ps, Known Length qs, Known Length ss, Known Length ts) => LearnFunc (Just (T ps)) (Just (T ss)) b c -> LearnFunc (Just (T qs)) (Just (T ts)) a b -> LearnFunc (Just (T (ps ++ qs))) (Just (T (ss ++ ts))) a c

Compose two LearnFunc on lists.

nilLF :: LearnFunc (Just (T '[])) (Just (T '[])) a a

Identity of .~

onlyLF :: forall p s a b. (KnownMayb p, MaybeC Num p, KnownMayb s, MaybeC Num s) => LearnFunc p s a b -> LearnFunc (Just (T (MaybeToList p))) (Just (T (MaybeToList s))) a b

LearnFunc with singleton lists; meant to be used with .~

data (:.~) :: Type -> Type -> Type where infixr 5

Compose two layers sequentially

Note that this composes in the opposite order of . and :.:, for consistency with the rest of the library.

The basic autoencoder is simply l :.~ m, where Learn a b l and Learn b a m. However, for sparse autoencoder, look at Autoencoder.

Constructors

(:.~) :: l -> m -> l :.~ m infixr 5

data (:&&&) :: Type -> Type -> Type where infixr 3

Fork/"Fan out" two models from the same input. Useful for

Constructors

(:&&&) :: l -> m -> l :&&& m infixr 3

type KAutoencoder n l m = l :.~ (FixedFunc (R n) (R n) :.~ m)

K-sparse autoencoder. A normal autoencoder is simply l :.~ m; however a k-sparse autoencoder attempts to ensure that the encoding has about k active components for every input.

http://www.ericlwilkinson.com/blog/2014/11/19/deep-learning-sparse-autoencoders

instance (Learn a (R n) l, Learn (R n) a m) => Learn a a (KAutoencoder n l m) where
    type LParamMaybe (KAutoencoder n l m) = TupMaybe (LParamMaybe l) (LParamMaybe m)
    type LStateMaybe (KAutoencoder n l m) = TupMaybe (LStateMaybe l) (LStateMaybe m)

To "encode" after training this, get the encoder with kaeEncoder.

kAutoencoder :: KnownNat n => l -> m -> Int -> KAutoencoder n l m

Construct a KAutoencoder.

Note that this only has a Learn instance of l outputs R n and m takes R n. Also, for this to be an actual autoencoder, the input of l has to be the same as the output of m.

kaeEncoder :: KAutoencoder n l m -> l

kaeDecoder :: KAutoencoder n l m -> m

data Feedback :: Type -> Type -> Type -> Type where

Take a model and turn it into a model that runs its output into itself several times, returning the final result. Parameterized by the number of repeats, and the function to process the output to become the next input.

I don't know why anyone would ever want this.

See FeedbackTrace if you want to observe all of the intermediate outputs.

Constructors

FB :: {..} -> Feedback a b l

Fields

_fbTimes :: Int

_fbFeed :: forall s. Reifies s W => BVar s b -> BVar s a

_fbLearn :: l

feedbackId :: Int -> l -> Feedback a a l

Construct a Feedback from an endofunction (a function that returns a value fo the same type as its input) by simply providing the output directly as the next intput.

data FeedbackTrace :: Nat -> Type -> Type -> Type -> Type where

Take a model and turn it into a model that runs its output into itself several times, and returns all of the intermediate outputs. Parameterized by the function to process the output to become the next input.

See Feedback if you only need the final result.

Compare also to Unroll.

Constructors

FBT :: {..} -> FeedbackTrace n a b l

Fields

_fbtFeed :: forall s. Reifies s W => BVar s b -> BVar s a

_fbtLearn :: l

feedbackTraceId :: l -> FeedbackTrace n a a l

Construct a FeedbackTrace from an endofunction (a function that returns a value fo the same type as its input) by simply providing the output directly as the next intput.