Backprop.Learn.Model.Function

Types

newtype ParamFunc p a b

A ParamFunc p a b is a parameterized function from a to b, potentially with trainable parameter p.

A utility wrapper for a deterministic and stateless model.

Constructors

PF

Fields

runParamFunc :: forall s. Reifies s W => Mayb (BVar s) p -> BVar s a -> BVar s b

type ParamFuncP p = ParamFunc (Just p)

Convenient type synonym for a ParamFunc with parameters.

Mostly made to be easy to construct/deconstruct with PFP and runParamFuncP.

pattern PFP :: (forall s. Reifies s W => BVar s p -> BVar s a -> BVar s b) -> ParamFuncP p a b

runParamFuncP :: ParamFuncP p a b -> forall s. Reifies s W => BVar s p -> BVar s a -> BVar s b

type FixedFunc = ParamFunc Nothing

An unparameterized function. Has a Category instance.

A FixedFunc a b essentially the same as a:

forall s. Reifies s W => BVar s a -> BVar s b

And the FF pattern (and runFixedFunc extractor) witness this.

It is usually better to just use the functions directly, with combinators like Dimap, LMap, RMap, dimapPF, lmapPF, rmapPF, etc. This is just provided to let you work nicely with ParamFunc combinators.

pattern FF :: (forall s. Reifies s W => BVar s a -> BVar s b) -> FixedFunc a b

FF and runFixedFunc witness the fact that a FixedFunc a b is just a function from BVar s a to BVar s b.

runFixedFunc :: FixedFunc a b -> forall s. Reifies s W => BVar s a -> BVar s b

paramMap :: KnownNat i => (forall s. Reifies s W => Mayb (BVar s) p -> BVar s Double -> BVar s Double) -> ParamFunc p (R i) (R i)

Utility function to make a ParamFunc that maps a parameterized function over every item in the R. The parameter is shared across the entire map, and trained.

learnParam :: forall l a b. (Learn a b l, NoState l) => l -> ParamFunc (LParamMaybe l) a b

Create a ParamFunc from any instance of Learn that does not have state.

idLearn :: FixedFunc a a

Identity model, useful for using with other combinators.

Combinators

dimapPF :: (forall s. Reifies s W => BVar s a -> BVar s b) -> (forall s. Reifies s W => BVar s c -> BVar s d) -> ParamFunc p b c -> ParamFunc p a d

Pre- and post-compose a ParamFunc

lmapPF :: (forall s. Reifies s W => BVar s a -> BVar s b) -> ParamFunc p b c -> ParamFunc p a c

Precompose a ParamFunc

rmapPF :: (forall s. Reifies s W => BVar s b -> BVar s c) -> ParamFunc p a b -> ParamFunc p a c

Postcompose a ParamFunc

compPF :: forall p q a b c. (MaybeC Num p, MaybeC Num q, KnownMayb p, KnownMayb q) => ParamFunc p a b -> ParamFunc q b c -> ParamFunc (TupMaybe p q) a c

Compose two ParamFuncs sequentially

parPF :: forall p q a b c d. (MaybeC Num p, MaybeC Num q, KnownMayb p, KnownMayb q, Num a, Num b, Num c, Num d) => ParamFunc p a c -> ParamFunc q b d -> ParamFunc (TupMaybe p q) (T2 a b) (T2 c d)

Compose two ParamFuncs in parallel

Chain

(.-) :: forall ps qs a b c. (ListC (Num <$> ps), ListC (Num <$> qs), Known Length ps, Known Length qs) => ParamFuncP (T ps) b c -> ParamFuncP (T qs) a b -> ParamFuncP (T (ps ++ qs)) a c infixr 9

Compose two ParamFuncPs on lists.

nilPF :: ParamFuncP (T '[]) a a

The identity of .-

onlyPF :: forall p a b. (KnownMayb p, MaybeC Num p) => ParamFunc p a b -> ParamFuncP (T (MaybeToList p)) a b

ParamFuncP taking a singleton list; meant to be used with .-

Activation functions

See https://en.wikipedia.org/wiki/Activation_function

* Maps

** Unparameterized

step :: (Ord a, Num a) => a -> a

Binary step / heaviside step

To use with vectors (R), use vmap'.

\[ \begin{cases} 0 & \text{for } x < 0 \\ 1 & \text{for } x \ge 0 \end{cases} \]

logistic :: Floating a => a -> a

Logistic function

\[ \sigma(x) = \frac{1}{1 + e^{-x}} \]

softsign :: Fractional a => a -> a

Softsign activation function

\[ \frac{x}{1 + \lvert x \rvert} \]

reLU :: (Num a, Ord a) => a -> a

Rectified linear unit.

To use with vectors (R), use vmap'.

\[ \max(0,x) \]

reLU = preLU 0

softPlus :: Floating a => a -> a

SoftPlus

\[ \ln(1 + e^x) \]

bentIdentity :: Floating a => a -> a

Bent identity

\[ \frac{\sqrt{x^2 + 1} - 1}{2} + x \]

siLU :: Floating a => a -> a

Sigmoid-weighted linear unit. Multiply logistic by its input.

\[ x \sigma(x) \]

softExponential

Arguments

:: (Floating a, Ord a)

=> a
α

-> a
x

-> a

SoftExponential

To use with vectors (R), use vmap'.

\[ \begin{cases} - \frac{\ln(1 - \alpha (x + \alpha))}{\alpha} & \text{for } \alpha < 0 \\ x & \text{for } \alpha = 0 \\ \frac{e^{\alpha x} - 1}{\alpha} + \alpha & \text{for } \alpha > 0 \end{cases} \]

sinc :: (Floating a, Eq a) => a -> a

Sinc

\[ \begin{cases} 1 & \text{for } x = 0 \\ \frac{\sin(x)}{x} & \text{for } x \ne 0 \end{cases} \]

gaussian :: Floating a => a -> a

Gaussian

\[ e^{-x^2} \]

tanh :: Floating a => a -> a

atan :: Floating a => a -> a

sin :: Floating a => a -> a

vmap :: (Reifies s W, KnownNat n) => (BVar s ℝ -> BVar s ℝ) -> BVar s (R n) -> BVar s (R n)

Note: if possible, use the potentially much more performant vmap'.

vmap' :: (Reifies s W, Num (vec n), Storable field, Sized field (vec n) Vector) => (forall s'. Reifies s' W => BVar s' field -> BVar s' field) -> BVar s (vec n) -> BVar s (vec n)

vmap, but potentially more performant. Only usable if the mapped function does not depend on any external BVars.

Parameterized

liftUniform :: (Reifies s W, KnownNat n) => (BVar s (R n) -> r) -> BVar s Double -> r

Usable with functions like *, isru, etc. to turn them into a form usable with PFP:

liftUniform (*)  :: BVar s Double -> BVar s (R n) -> BVar s (R n)
liftUniform isru :: BVar s Double -> BVar s (R n) -> BVar s (R n)

Basically turns a parmaeterized function on individual elements of into one that shares the same parameter across all elements of the vector.

isru

Arguments

:: Floating a

=> a
α (scaling parameter)

-> a
x

-> a

Inverse square root unit

\[ \frac{x}{\sqrt{1 + \alpha x^2}} \]

See liftUniform to make this compatible with PFP.

You can also just use this after partially applying it, to fix the parameter (and not have it trained).

preLU

Arguments

:: (Num a, Ord a)

=> a
α (scaling parameter)

-> a
x

-> a

Parametric rectified linear unit

To use with vectors (R), use vmap'.

If scaling parameter is a fixed (and not learned) parameter, this is typically called a leaky recitified linear unit (typically with α = 0.01).

To use as a learned parameter:

vmap . preLU :: BVar s Double -> BVar s (R n) -> BVar s (R n)

This can be give directly to PFP.

To fix the paramater ("leaky"), just partially apply a parameter:

preLU 0.01           :: BVar s (R n) -> BVar s (R n)
preLU (realToFrac α) :: BVar s (R n) -> BVar s (R n)

See also rreLU.

\[ \begin{cases} \alpha x & \text{for } x < 0 \\ x & \text{for } x \ge 0 \end{cases} \]

sreLU

Arguments

:: (Num a, Ord a)

=> a
t_l

-> a
a_l

-> a
t_r

-> a
a_l

-> a
x

-> a

S-shaped rectified linear activiation unit

See sreLUPFP for an uncurried and uniformly lifted version usable with PFP.

\[ \begin{cases} t_l + a_l (x - t_l) & \text{for } x \le t_l \\ x & \text{for } t_l < x < t_r \\ t_r + a_r (x - t_r) & \text{for } x \ge t_r \end{cases} \]

sreLUPFP :: (KnownNat n, Reifies s W) => BVar s (T2 (T2 Double Double) (T2 Double Double)) -> BVar s (R n) -> BVar s (R n)

An uncurried and uniformly lifted version of sreLU directly usable with PFP.

eLU

Arguments

:: (Floating a, Ord a)

=> a
α (scaling parameter)

-> a
x

-> a

Exponential linear unit

To use as a learned parameter:

vmap . eLU :: BVar s Double -> BVar s (R n) -> BVar s (R n)

This can be give directly to PFP.

To fix the paramater, just partially apply a parameter:

eLU 0.01 :: BVar s (R n) -> BVar s (R n)

\[ \begin{cases} \alpha (e^x - 1) & \text{for } x < 0 \\ x & \text{for } x \ge 0 \end{cases} \]

isrLU

Arguments

:: (Floating a, Ord a)

=> a
α

-> a
x

-> a

Inverse square root linear unit

To use with vectors (R), use vmap'.

\[ \begin{cases} \frac{x}{1 + \alpha x^2} & \text{for } x < 0 \\ x & \text{for } x \ge 0 \end{cases} \]

apl

Arguments

:: (KnownNat n, KnownNat m, Reifies s W)

=> BVar s (L n m)
a

-> BVar s (L n m)
b

-> BVar s (R m)
x

-> BVar s (R m)

Adaptive piecewise linear activation unit

See aplPFP for an uncurried version usable with PFP.

\[ \max(0, x_i) + \sum_j^M a_i^j \max(0, -x_i + b_i^j) \]

aplPFP :: (KnownNat n, KnownNat m, Reifies s W) => BVar s (T2 (L n m) (L n m)) -> BVar s (R m) -> BVar s (R m)

apl uncurried, to be directly usable with PFP.

Mixing

softMax :: (KnownNat i, Reifies s W) => BVar s (R i) -> BVar s (R i)

Softmax normalizer

maxout :: (KnownNat n, Reifies s W) => BVar s (R n) -> BVar s Double

Maximum of vector.

Compare to norm_InfV, which gives the maximum absolute value.

kSparse

Arguments

:: (Reifies s W, KnownNat n)

=> Int
number of items to keep

-> BVar s (R n)

-> BVar s (R n)

Keep only the top k values, and zero out all of the rest.

Useful for postcomposing in between layers (with a logistic function before) to encourage the number of "activated" neurons is kept to be around k. Used in k-Sprase autoencoders (see KAutoencoder).

http://www.ericlwilkinson.com/blog/2014/11/19/deep-learning-sparse-autoencoders