Backprop.Learn.Model.Neural

Fully connected

Feed-forward

type FC i o = LinReg i o

Fully connected feed-forward layer with bias. Parameterized by its initialization distribution.

Note that this has no activation function; to use as a model with activation function, chain it with an activation function using RMap, :.~, etc.; see FCA for a convenient type synonym and constructor.

Without any activation function, this is essentially a multivariate linear regression.

With the logistic function as an activation function, this is essentially multivariate logistic regression. (See Logistic)

fc :: ContGen d => d -> FC i o

Construct an FC i o using a given distribution from the statistics library.

type FCp = LRp

fcBias :: Lens' (FCp i o) (R o)

fcWeights :: Lens (FCp i o) (FCp i' o) (L o i) (L o i')

With activation function

type FCA i o = RMap (R o) (R o) (FC i o)

Convenient synonym for an FC post-composed with a simple parameterless activation function.

pattern FCA :: (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCA i o

Construct an FCA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

fca :: ContGen d => d -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCA i o

Construct an FCA i o using a given distribution from the statistics library.

_fcaGen :: FCA i o -> forall (m :: * -> *). PrimMonad m => Gen (PrimState m) -> m Double

_fcaActivation :: FCA i o -> forall s. Reifies s W => BVar s (R o) -> BVar s (R o)

Recurrent

type FCR h i o = Recurrent (R (i + h)) (R i) (R h) (R o) (FC (i + h) o)

Fully connected recurrent layer with bias.

Parameterized by its initialization distributions, and also the function to compute the new state from previous input.

pattern FCR :: (KnownNat h, KnownNat i) => (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall m. PrimMonad m => Gen (PrimState m) -> m (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> FCR h i o

fcr

Arguments

:: (KnownNat h, KnownNat i, ContGen d, ContGen e)

=> d
gen params

-> e
gen history

-> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h))

-> FCR h i o

Construct an FCR h i o using given distributions from the statistics library.

type FCRp h i o = FCp (i + h) o

fcrBias :: Lens' (FCRp h i o) (R o)

fcrInputWeights :: (KnownNat h, KnownNat i, KnownNat i', KnownNat o) => Lens (FCRp h i o) (FCRp h i' o) (L o i) (L o i')

fcrStateWeights :: (KnownNat h, KnownNat h', KnownNat i, KnownNat o) => Lens (FCRp h i o) (FCRp h' i o) (L o h) (L o h')

With activation function

type FCRA h i o = RMap (R o) (R o) (FCR h i o)

Convenient synonym for an FCR post-composed with a simple parameterless activation function.

pattern FCRA :: (KnownNat h, KnownNat i) => (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall m. PrimMonad m => Gen (PrimState m) -> m (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCRA h i o

Construct an FCRA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

fcra

Arguments

:: (KnownNat h, KnownNat i, ContGen d, ContGen e)

=> d
gen params

-> e
gen state

-> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h))

-> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o))

-> FCRA h i o

Construct an FCRA h i o using given distributions from the statistics library.

_fcraGen :: FCRA h i o -> forall (m :: * -> *). (PrimMonad m, KnownNat h, KnownNat i) => Gen (PrimState m) -> m Double

_fcraGenState :: FCRA h i o -> forall (m :: * -> *). (PrimMonad m, KnownNat h, KnownNat i) => Gen (PrimState m) -> m (R h)

_fcraStore :: FCRA h i o -> forall s. (Reifies s W, KnownNat h, KnownNat i) => BVar s (R o) -> BVar s (R h)

_fcraActivation :: FCRA h i o -> forall s. (Reifies s W, KnownNat h, KnownNat i) => BVar s (R o) -> BVar s (R o)