Backprop.Learn.Model.Neural.LSTM

Basic LSTM

type LSTM i o = Recurrent (R (i + o)) (R i) (R o) (R o) (LSTM' i o)

Long-term short-term memory layer

http://colah.github.io/posts/2015-08-Understanding-LSTMs/

instance Learn (R i) (R o) (LSTM i o) where
    type LParamMaybe (LSTM i o) = 'Just (LSTMp i o)
    type LStateMaybe (LSTM i o) = 'Just (T2 (R o) (R o))     -- cell state, hidden state

pattern LSTM :: (KnownNat i, KnownNat o) => LSTM i o

Construct an LSTM

data LSTMp (i :: Nat) (o :: Nat)

LSTM layer parmateters

Constructors

LSTMp

Fields

_lstmForget :: !(FCp (i + o) o)

_lstmInput :: !(FCp (i + o) o)

_lstmUpdate :: !(FCp (i + o) o)

_lstmOutput :: !(FCp (i + o) o)

lstmForget :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmInput :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmUpdate :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmOutput :: Lens' (LSTMp i o) (FCp (i + o) o)

data LSTM' (i :: Nat) (o :: Nat)

Base for LSTM. An LSTM is just an LSTM' where the second half of the input vector is the previous output.

This is mostly an implementation detail. It is recommended that you use LSTM and its constructor, instead of directly using this.

Constructors

LSTM'

GRU

type GRU i o = Recurrent (R (i + o)) (R i) (R o) (R o) (GRU' i o)

Gated Recurrent Unit

http://colah.github.io/posts/2015-08-Understanding-LSTMs/

instance Learn (R i) (R o) (GRU i o) where
    type LParamMaybe (GRU i o) = 'Just (GRUp i o)
    type LStateMaybe (GRU i o) = 'Just (R o)

pattern GRU :: (KnownNat i, KnownNat o) => GRU i o

Construct an GRU

data GRUp (i :: Nat) (o :: Nat)

GRU layer parmateters

Constructors

GRUp

Fields

_gruMemory :: !(FCp (i + o) o)

_gruUpdate :: !(FCp (i + o) o)

_gruOutput :: !(FCp (i + o) o)

gruMemory :: Lens' (GRUp i o) (FCp (i + o) o)

gruUpdate :: Lens' (GRUp i o) (FCp (i + o) o)

gruOutput :: Lens' (GRUp i o) (FCp (i + o) o)

data GRU' (i :: Nat) (o :: Nat)

Base for GRU. An GRU is just an GRU' where the second half of the input vector is the previous output.

This is mostly an implementation detail. It is recommended that you use GRU and its constructor, instead of directly using this.

Constructors

GRU'