Backprop.Learn.Model.Parameter

Documentation

deParam

Arguments

:: (Backprop p, Backprop q, Backprop pq)

=> (pq -> (p, q))
split

-> (p -> q -> pq)
join

-> q
fixed param

-> (forall m. PrimMonad m => Gen (PrimState m) -> m q)
fixed stoch param

-> Model (Just pq) s a b

-> Model (Just p) s a b

Fix a part of a parameter as constant, preventing backpropagation through it and not training it.

Treats a pq parameter as essentialyl a (p, q), witnessed through the split and join functions.

Takes the fixed value of q, as well as a stochastic mode version with fixed distribution.

deParamD

Arguments

:: (Backprop p, Backprop q, Backprop pq)

=> (pq -> (p, q))
split

-> (p -> q -> pq)
join

-> q
fixed param

-> Model (Just pq) s a b

-> Model (Just p) s a b

deParam, but with no special stochastic mode version.

reParam :: (forall z. Reifies z W => Mayb (BVar z) q -> Mayb (BVar z) p) -> (forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> Mayb (BVar z) q -> m (Mayb (BVar z) p)) -> Model p s a b -> Model q s a b

Pre-applies a function to a parameter before a model sees it. Essentially something like lmap for parameters.

Takes a determinstic function and also a stochastic function for stochastic mode.

reParamD :: (forall z. Reifies z W => Mayb (BVar z) q -> Mayb (BVar z) p) -> Model p s a b -> Model q s a b

reParam, but with no special stochastic mode function.

dummyParam :: Model Nothing s a b -> Model p s a b

Give an unparameterized model a "dummy" parameter.