Backprop.Learn.Model.Regression

Documentation

type Linear n m = FC n m

Multivariate linear regression, from an n-vector to an m-vector.

Is essentially just a fully connected neural network layer with bias, with no activation function.

linearRegression :: (forall f. PrimMonad f => Gen (PrimState f) -> f Double) -> Linear n m

Construct a linear regression model from an initialization function for coefficients

type Logistic n = RMap (R n) (R 1) Double (Linear n 1)

logisticRegression :: (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> Logistic n

Construct a logistic regression model from an initialization function for coefficients.