Backprop.Learn.Model.State

Make models stateless

newtype TrainState :: Type -> Type where

Make a model stateless by converting the state to a trained parameter, and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See DeState, as well.

Its parameters are:

If l has no parameters, just the initial state.

If l has parameters, a T2 of the parameter and initial state.

Constructors

TrainState :: {..} -> TrainState l

Fields

getTrainState :: l

data DeState :: Type -> Type -> Type where

Make a model stateless by pre-applying a fixed state (or a stochastic one with fixed stribution) and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See TrainState, as well.

Constructors

DS :: {..} -> DeState s l

Fields

_dsInitState :: s

_dsInitStateStoch :: forall m. PrimMonad m => Gen (PrimState m) -> m s

_dsLearn :: l

dsDeterm :: s -> l -> DeState s l

Create a DeState from a deterministic, non-stochastic initialization function.

Manipulate model states

Unroll

newtype Unroll :: Nat -> Type -> Type where

Unroll a (usually) stateful model into one taking a vector of sequential inputs.

Basically applies the model to every item of input and returns all of the results, but propagating the state between every step.

Useful when used before TrainState or DeState. See UnrollTrainState and UnrollDeState.

Compare to FeedbackTrace, which, instead of receiving a vector of sequential inputs, receives a single input and uses its output as the next input.

Constructors

Unroll :: {..} -> Unroll n l

Fields

getUnroll :: l

type UnrollTrainState n l = TrainState (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and treat the initial state as a trained parameter.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollTrainState n l)

    type LParamMaybe (UnrollDeState n l) = TupMaybe (LParamMaybe l) (LStateMaybe l)
    type LStateMaybe (UnrollDeState n l) = 'Nothing

pattern UnrollTrainState :: l -> UnrollTrainState n l

Constructor and pattern for UnrollTrainState

getUnrollTrainState :: UnrollTrainState n l -> l

type UnrollDeState n l = DeState (LState l) (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and fix the initial state.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollDeState n l)

    type LParamMaybe (UnrollDeState n l) = LParamMaybe l
    type LStateMaybe (UnrollDeState n l) = 'Nothing

pattern UDS :: LState l -> (forall m. PrimMonad m => Gen (PrimState m) -> m (LState l)) -> l -> UnrollDeState n l

Constructor and pattern for UnrollDeState

_udsInitState :: UnrollDeState n l -> LState l

_udsInitStateStoch :: UnrollDeState n l -> forall (m :: * -> *). PrimMonad m => Gen (PrimState m) -> m (LState l)

_udsLearn :: UnrollDeState n l -> l

ReState

data ReState :: Type -> Type -> Type -> Type where

Transform the state of a model by providing functions to pre-apply and post-apply before and after the original model sees the state.

I don't really know why you would ever want to do this, but it was fun to write.

ReState p '()' is essentially DeState.

Constructors

RS :: {..} -> ReState s t l

Fields

_rsInitState :: forall m. PrimMonad m => Gen (PrimState m) -> m t

_rsTo :: forall q. Reifies q W => BVar q s -> BVar q t

_rsFrom :: forall q. Reifies q W => BVar q t -> BVar q s

_rsFromStoch :: forall m q. (PrimMonad m, Reifies q W) => Gen (PrimState m) -> BVar q t -> m (BVar q s)

_rsLearn :: l

rsDeterm

Arguments

:: (forall m. PrimMonad m => Gen (PrimState m) -> m t)
initialize state

-> (forall q. Reifies q W => BVar q s -> BVar q t)
to

-> (forall q. Reifies q W => BVar q t -> BVar q s)
from

-> l
model

-> ReState s t l

Create a ReState from a deterministic, non-stochastic transformation function.