Backprop.Learn.Model.Types

Model type

type ModelFunc p s a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> (BVar z b, PMaybe (BVar z) s)

type ModelFuncStoch p s a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> m (BVar z b, PMaybe (BVar z) s)

data Model :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

General parameterized model with potential state

Constructors

Model

Fields

:: { runLearn :: ModelFunc p s a b

, runLearnStoch :: ModelFuncStoch p s a b

} -> Model p s a b

modelD :: ModelFunc p s a b -> Model p s a b

Construct a deterministic model, with no stochastic component.

Specialized Models

Stateless

type ModelFuncStateless p a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> BVar z b

type ModelFuncStochStateless p a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> m (BVar z b)

type ModelStateless p = Model p Nothing

Parameterized model with no state

pattern ModelStateless :: ModelFuncStateless p a b -> ModelFuncStochStateless p a b -> ModelStateless p a b

Construct a ModelStateless.

runLearnStateless :: ModelStateless p a b -> forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> BVar z b

runLearnStochStateless :: ModelStateless p a b -> forall (m :: Type -> Type) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> m (BVar z b)

modelStatelessD :: ModelFuncStateless p a b -> ModelStateless p a b

Construct a deterministic stateless model, with no stochastic component.

Stateless and Parameterless

type BFunc a b = forall z. Reifies z W => BVar z a -> BVar z b

type BFuncStoch a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

type Func = Model Nothing Nothing

pattern Func :: BFunc a b -> BFuncStoch a b -> Func a b

Unparameterized model with no state

runFunc :: Func a b -> forall z. Reifies z W => BVar z a -> BVar z b

runFuncStoch :: Func a b -> forall (m :: Type -> Type) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

funcD :: BFunc a b -> Func a b

Construct an deterministic unparameterized stateless model, with no stochastic component.

Manipulating models as functions

type ModelFuncM m p s a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> m (BVar z b, PMaybe (BVar z) s)

withModelFunc0 :: (forall m. Monad m => ModelFuncM m p s a b) -> Model p s a b

withModelFunc :: (forall m. Monad m => ModelFuncM m p s a b -> ModelFuncM m q t c d) -> Model p s a b -> Model q t c d

withModelFunc2 :: (forall m. Monad m => ModelFuncM m p s a b -> ModelFuncM m q t c d -> ModelFuncM m r u e f) -> Model p s a b -> Model q t c d -> Model r u e f

Utility

data PMaybe (a :: k -> Type) (b :: Maybe k) :: forall k. (k -> Type) -> Maybe k -> Type where

A PMaybe f 'Nothing contains nothing, and a PMaybe f ('Just a) contains an f a.

In practice this can be useful to write polymorphic functions/abstractions that contain an argument that can be "turned off" for different instances.

Constructors

PNothing :: forall k (a :: k -> Type) (b :: Maybe k). PMaybe a (Nothing :: Maybe k)

PJust :: forall k (a :: k -> Type) (b :: Maybe k) (a1 :: k). a a1 -> PMaybe a (Just a1)

Bundled Patterns

pattern TJust :: a -> TMaybe (Just a)

type TMaybe = PMaybe TF

fromPJust :: PMaybe f (Just a) -> f a

type Learnables as = (RecApplicative as, ReifyConstraint Backprop TF as, RMap as, RApply as)

Combination of common constraints for type-level lists.