Backprop.Learn.Regularize

Documentation

type Regularizer p = forall s. Reifies s W => BVar s p -> BVar s Double

A regularizer on parameters

class Backprop p => Regularize p where

A class for data types that support regularization during training.

This class is somewhat similar to Metric Double, in that it supports summing the components and summing squared components. However, the main difference is that when summing components, we only consider components that we want to regularize.

Often, this doesn't include bias terms (terms that "add" to inputs), and only includes terms that "scale" inputs, like components in a weight matrix of a feed-forward neural network layer.

However, if all of your components are to be regularized, you can use norm_1, norm_2, lassoLinear, and ridgeLinear as sensible implementations, or use DerivingVia with RegularizeMetric:

data MyType = ...
  deriving Regularize via (RegularizeMetric MyType)

You can also derive an instance where no are regularized, using NoRegularize:

data MyType = ...
  deriving Regularize via (NoRegularize MyType)

The default implementations are based on Generics, and work for types that are records of items that are all instances of Regularize.

Minimal complete definition

Nothing

Methods

rnorm_1 :: p -> Double

Like norm_1: sums all of the weights in p, but only the ones you want to regularize:

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_1 = norm_1. However, this would count all terms in p, even potential bias terms.

rnorm_1 :: (ADT p, Constraints p Regularize) => p -> Double

Like norm_1: sums all of the weights in p, but only the ones you want to regularize:

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_1 = norm_1. However, this would count all terms in p, even potential bias terms.

rnorm_2 :: p -> Double

Like norm_2: sums all of the squares of the weights n p, but only the ones you want to regularize:

\[ \sum_w w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_2 = norm_2. However, this would count all terms in p, even potential bias terms.

rnorm_2 :: (ADT p, Constraints p Regularize) => p -> Double

Like norm_2: sums all of the squares of the weights n p, but only the ones you want to regularize:

\[ \sum_w w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_2 = norm_2. However, this would count all terms in p, even potential bias terms.

lasso :: Double -> p -> p

lasso r p sets all regularized components (that is, components summed by rnorm_1) in p to be either r if that component was positive, or -r if that component was negative. Behavior is not defined if the component is exactly zero, but either r or -r are sensible possibilities.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_1 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set lasso = lassoLinear. However, this is only valid if rnorm_1 counts all terms in p, including potential bias terms.

lasso :: (ADT p, Constraints p Regularize) => Double -> p -> p

lasso r p sets all regularized components (that is, components summed by rnorm_1) in p to be either r if that component was positive, or -r if that component was negative. Behavior is not defined if the component is exactly zero, but either r or -r are sensible possibilities.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_1 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set lasso = lassoLinear. However, this is only valid if rnorm_1 counts all terms in p, including potential bias terms.

ridge :: Double -> p -> p

ridge r p scales all regularized components (that is, components summed by rnorm_2) in p by r.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_2 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set ridge = ridgeLinear. However, this is only valid if rnorm_2 counts all terms in p, including potential bias terms.

ridge :: (ADT p, Constraints p Regularize) => Double -> p -> p

ridge r p scales all regularized components (that is, components summed by rnorm_2) in p by r.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_2 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set ridge = ridgeLinear. However, this is only valid if rnorm_2 counts all terms in p, including potential bias terms.

l1Reg :: Regularize p => Double -> Regularizer p

Backpropagatable L1 regularization; also known as lasso regularization.

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

l2Reg :: Regularize p => Double -> Regularizer p

Backpropagatable L2 regularization; also known as ridge regularization.

\[ \sum_w w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

noReg :: Regularizer p

No regularization

l2RegMetric

Arguments

:: (Metric Double p, Backprop p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

L2 regularization for instances of Metric. This will count all terms, including any potential bias terms.

This is what l2Reg would be for a type p if you declare an instance of Regularize with rnorm_2 = norm_2, and ridge = ridgeLinear.

l1RegMetric

Arguments

:: (Num p, Metric Double p, Backprop p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

L1 regularization for instances of Metric. This will count all terms, including any potential bias terms.

This is what l1Reg would be for a type p if you declare an instance of Regularize with rnorm_1 = norm_1, and lasso = lassoLinear.

Build instances

DerivingVia

newtype RegularizeMetric a

Newtype wrapper (meant to be used with DerivingVia) to derive an instance of Regularize that uses its Metric instance, and regularizes every component of a data type, including any potential bias terms.

Constructors

RegularizeMetric a

newtype NoRegularize a

Newtype wrapper (meant to be used with DerivingVia) to derive an instance of Regularize that does not regularize any part of the type.

Constructors

NoRegularize a

Linear

lassoLinear :: (Linear Double p, Num p) => Double -> p -> p

A default implementation of lasso for instances of Linear Double and Num. However, this is only valid if the corresponding rnorm_1 counts all terms in p, including potential bias terms.

ridgeLinear :: Linear Double p => Double -> p -> p

A default implementation of ridge for instances of Linear Double. However, this is only valid if the corresponding rnorm_2 counts all terms in p, including potential bias terms.

Generics

grnorm_1 :: (ADT p, Constraints p Regularize) => p -> Double

grnorm_2 :: (ADT p, Constraints p Regularize) => p -> Double

glasso :: (ADT p, Constraints p Regularize) => Double -> p -> p

gridge :: (ADT p, Constraints p Regularize) => Double -> p -> p

Manipulate regularizers

addReg :: Regularizer p -> Regularizer p -> Regularizer p

Add together two regularizers

scaleReg :: Double -> Regularizer p -> Regularizer p

Scale a regularizer's influence