Backprop.Learn.Train

Gradients

gradModelLoss :: Backprop p => Loss b -> Regularizer p -> Model (Just p) Nothing a b -> p -> a -> b -> p

Gradient of model with respect to loss function and target

gradModelStochLoss :: (Backprop p, PrimMonad m) => Loss b -> Regularizer p -> Model (Just p) Nothing a b -> Gen (PrimState m) -> p -> a -> b -> m p

Stochastic gradient of model with respect to loss function and target

Opto

type Grad (m :: * -> *) a = a -> m (Diff a)

modelGrad :: (MonadSample (a, b) m, Backprop p) => Loss b -> Regularizer p -> Model (Just p) Nothing a b -> Grad m p

Using a model's deterministic prediction function (with a given loss function), generate a Grad compatible with Numeric.Opto and Numeric.Opto.Run.

modelGradStoch :: (MonadSample (a, b) m, PrimMonad m, Backprop p) => Loss b -> Regularizer p -> Model (Just p) Nothing a b -> Gen (PrimState m) -> Grad m p

Using a model's stochastic prediction function (with a given loss function), generate a Grad compatible with Numeric.Opto and Numeric.Opto.Run.