Learn.Neural.Network

Documentation

data LChain :: Type where

Constructors

(:~) :: [Nat] -> Type -> LChain

type (:~) s c = s :~ c

data Network :: RunMode -> ([Nat] -> Type) -> LChain -> [LChain] -> [Nat] -> Type where

Constructors

NetExt :: (ComponentLayer r c b i o, CConstr c b i o) => Layer r c b i o -> Network r b (i :~ c) '[] o

(:&) :: (ComponentLayer r c b i h, CConstr c b i h, Num (b h)) => Layer r c b i h -> Network r b (h :~ d) hs o -> Network r b (i :~ c) ((h :~ d) ': hs) o

netOp :: forall b i c hs o r s. (BLAS b, Num (b i), Num (b o)) => OpB s '[Network r b (i :~ c) hs o, b i] '[Network r b (i :~ c) hs o, b o]

netOpPure :: forall b i c hs o s. (BLAS b, Num (b i), Num (b o)) => OpB s '[Network FeedForward b (i :~ c) hs o, b i] '[b o]

netOpRecurrent :: forall n b i c hs o s. (Known (NetStruct Recurrent b (i :~ c) hs) o, BLAS b, Num (b o), Num (b i)) => Nat n -> OpB s (Network Recurrent b (i :~ c) hs o ': Replicate n (b i)) (Network Recurrent b (i :~ c) hs o ': Replicate n (b o))

netOpRecurrent_ :: forall n b i c hs o s. (Known (NetStruct Recurrent b (i :~ c) hs) o, BLAS b, Num (b o), Num (b i)) => Nat n -> OpB s (Network Recurrent b (i :~ c) hs o ': Replicate n (b i)) (Replicate n (b o))

netOpRecurrentLast :: forall n b i c hs o s. (Known (NetStruct Recurrent b (i :~ c) hs) o, BLAS b, Num (b o), Num (b i)) => Nat n -> OpB s (Network Recurrent b (i :~ c) hs o ': (b i ': Replicate n (b i))) '[Network Recurrent b (i :~ c) hs o, b o]

netOpRecurrentLast_ :: forall n b i c hs o s. (Known (NetStruct Recurrent b (i :~ c) hs) o, BLAS b, Num (b o), Num (b i)) => Nat n -> OpB s (Network Recurrent b (i :~ c) hs o ': (b i ': Replicate n (b i))) '[b o]

runNet :: (Num (b i), Num (b o), BLAS b) => Network r b (i :~ c) hs o -> b i -> (b o, Network r b (i :~ c) hs o)

runNetPure :: (Num (b i), Num (b o), BLAS b) => Network FeedForward b (i :~ c) hs o -> b i -> b o

runNetRecurrent :: (Num (b i), Num (b o), BLAS b) => Network Recurrent b (i :~ c) hs o -> [b i] -> ([b o], Network Recurrent b (i :~ c) hs o)

runNetRecurrentLast :: (Num (b i), Num (b o), BLAS b) => Network Recurrent b (i :~ c) hs o -> NonEmpty (b i) -> (b o, Network Recurrent b (i :~ c) hs o)

runNetFeedback :: (Num (b i), Num (b o), BLAS b) => Nat n -> (b o -> (a, b i)) -> Network Recurrent b (i :~ c) hs o -> b i -> (Vec n a, Network Recurrent b (i :~ c) hs o)

runNetFeedbackM :: (Num (b i), Num (b o), BLAS b, Monad m) => Nat n -> (b o -> m (a, b i)) -> Network Recurrent b (i :~ c) hs o -> b i -> m (Vec n a, Network Recurrent b (i :~ c) hs o)

runNetFeedbackForever :: (Num (b i), Num (b o), BLAS b) => (b o -> (a, b i)) -> Network Recurrent b (i :~ c) hs o -> b i -> [a]

data NetConf :: RunMode -> ([Nat] -> Type) -> LChain -> [LChain] -> [Nat] -> Type where

Constructors

NCExt :: (ComponentLayer r c b i o, CConstr c b i o) => CConf c b i o -> NetConf r b (i :~ c) '[] o

(:&~) :: (ComponentLayer r c b i h, CConstr c b i h, SingI h, Num (b h)) => CConf c b i h -> NetConf r b (h :~ d) hs o -> NetConf r b (i :~ c) ((h :~ d) ': hs) o

initNet :: forall b i c hs o m r. (PrimMonad m, BLAS b, SingI i, SingI o) => NetConf r b (i :~ c) hs o -> Gen (PrimState m) -> m (Network r b (i :~ c) hs o)

data NetStruct :: RunMode -> ([Nat] -> Type) -> LChain -> [LChain] -> [Nat] -> Type where

Constructors

NSExt :: (ComponentLayer r c b i o, CConstr c b i o) => NetStruct r b (i :~ c) '[] o

NSInt :: (SingI h, Num (b h), ComponentLayer r c b i h, CConstr c b i h) => NetStruct r b (h :~ d) hs o -> NetStruct r b (i :~ c) ((h :~ d) ': hs) o

defNetConf' :: NetStruct r b i hs o -> NetConf r b i hs o

defNetConf :: Known (NetStruct r b i hs) o => NetConf r b i hs o

initDefNet' :: forall b i c hs o m r. (PrimMonad m, BLAS b, SingI i, SingI o) => NetStruct r b (i :~ c) hs o -> Gen (PrimState m) -> m (Network r b (i :~ c) hs o)

initDefNet :: (PrimMonad m, BLAS b, SingI i, SingI o, Known (NetStruct r b (i :~ c) hs) o) => Gen (PrimState m) -> m (Network r b (i :~ c) hs o)

data SomeNet :: RunMode -> ([Nat] -> Type) -> [Nat] -> [Nat] -> Type where

Constructors

SomeNet :: NetStruct r b (i :~ c) hs o -> Network r b (i :~ c) hs o -> SomeNet r b i o

someNet :: Known (NetStruct r b (i :~ c) hs) o => Network r b (i :~ c) hs o -> SomeNet r b i o

takeNet :: forall i c hs h d js o r b. NetStruct r b (i :~ c) hs h -> Network r b (i :~ c) (hs ++ ((h :~ d) ': js)) o -> Network r b (i :~ c) hs h

dropNet :: forall i c hs h d js o r b. NetStruct r b (i :~ c) hs h -> Network r b (i :~ c) (hs ++ ((h :~ d) ': js)) o -> Network r b (h :~ d) js o

splitNet :: forall i c hs h d js o r b. NetStruct r b (i :~ c) hs h -> Network r b (i :~ c) (hs ++ ((h :~ d) ': js)) o -> (Network r b (i :~ c) hs h, Network r b (h :~ d) js o)

traceNet :: forall r b i c hs o. (Num (b i), Num (b o), BLAS b) => Network r b (i :~ c) hs o -> b i -> ((Prod b (NetTrace hs), b o), Network r b (i :~ c) hs o)

traceNetPure :: forall b i c hs o. (Num (b i), Num (b o), BLAS b) => Network FeedForward b (i :~ c) hs o -> b i -> (Prod b (NetTrace hs), b o)