Learn.Neural.Train

Documentation

data OptimizerM :: (Type -> Type) -> Type -> Type -> Type where

Constructors

MkOM :: {..} -> OptimizerM m a p

Fields

oState :: s

oGrad :: a -> p -> m p

oUpdate :: p -> p -> s -> m (p, s)

params, grad, state

type Optimizer = OptimizerM I

pattern MkO :: forall a p. () => forall s. s -> (a -> p -> p) -> (p -> p -> s -> (p, s)) -> Optimizer a p

runOptimizer :: a -> p -> Optimizer a p -> (p, Optimizer a p)

runOptimizerM :: Monad m => a -> p -> OptimizerM m a p -> m (p, OptimizerM m a p)

runOptimizer_ :: a -> p -> Optimizer a p -> p

runOptimizerM_ :: Monad m => a -> p -> OptimizerM m a p -> m p

optimizeList :: forall a p. [a] -> p -> Optimizer a p -> (p, Optimizer a p)

optimizeListM :: forall m a p. Monad m => [a] -> p -> OptimizerM m a p -> m (p, OptimizerM m a p)

optimizeList_ :: [a] -> p -> Optimizer a p -> p

optimizeListM_ :: Monad m => [a] -> p -> OptimizerM m a p -> m p

optimizeListBatches :: [a] -> p -> Optimizer [a] p -> Int -> (p, Optimizer [a] p)

optimizeListBatchesM :: Monad m => [a] -> p -> OptimizerM m [a] p -> Int -> m (p, OptimizerM m [a] p)

optimizeListBatches_ :: [a] -> p -> Optimizer [a] p -> Int -> p

optimizeListBatchesM_ :: Monad m => [a] -> p -> OptimizerM m [a] p -> Int -> m p

batching :: forall f m a p. (Foldable f, Fractional p, Monad m) => OptimizerM m a p -> OptimizerM m (f a) p

sgdOptimizer :: forall p as bs c. (Fractional p, Num c, Every Num as, Every Num bs, Known Length as) => Double -> (forall s. OpB s (p ': as) bs) -> LossFunction bs c -> Optimizer (Tuple as, Tuple bs) p

sgdOptimizerM :: forall m p as bs c. (Fractional p, Num c, Every Num as, Every Num bs, Known Length as, Monad m) => Double -> m (OpBS (p ': as) bs) -> LossFunction bs c -> OptimizerM m (Tuple as, Tuple bs) p

adamOptimizer :: forall p as bs c. (Floating p, Num c, Every Num as, Every Num bs, Known Length as) => Adam -> (forall s. OpB s (p ': as) bs) -> LossFunction bs c -> Optimizer (Tuple as, Tuple bs) p

adamOptimizerM :: forall m p as bs c. (Floating p, Num c, Every Num as, Every Num bs, Known Length as, Monad m) => Adam -> m (OpBS (p ': as) bs) -> LossFunction bs c -> OptimizerM m (Tuple as, Tuple bs) p

data Adam

Constructors

Adam

Fields

adamStep :: Double

adamDecay1 :: Double

adamDecay2 :: Double

adamEpsilon :: Double

slidingParts :: Nat n -> [a] -> [Vec n a]

slidingPartsSplit :: Nat n -> [(a, b)] -> [(Vec n a, Vec n b)]

slidingPartsLast :: Nat (S n) -> [(a, b)] -> [(Vec (S n) a, b)]

asFeedback :: [a] -> [(a, a)]

class Default a where

A class for types with a default value.

Methods

def :: a

The default value for this type.