Backprop.Learn.Class

Documentation

class (Num p, Num a, Num b) => Learn p a b l | l -> p, l -> a, l -> b where

Class for models that can be trained using gradient descent.

An instance l of Learn p a b is parameterized by p, takes a as input, and returns b as outputs.

l can be thought of as representing the hyperparameters of the model, with p representing the trained parameters of the model.

If no trained parameters exist, NoParam can be used. This will automatically derive initParam.

Minimal complete definition

runFixed

Methods

initParam :: PrimMonad m => l -> Gen (PrimState m) -> m p

Initialization of trainiable model parameters.

initParam :: (PrimMonad m, p ~ NoParam) => l -> Gen (PrimState m) -> m p

Initialization of trainiable model parameters.

runFixed :: Reifies s W => l -> BVar s p -> BVar s a -> BVar s b

Run the model.

runStoch :: (PrimMonad m, Reifies s W) => l -> Gen (PrimState m) -> BVar s p -> BVar s a -> m (BVar s b)

Running the model in "stochastic" mode.

For nonstochastic models, is automatically defined in terms of runFixed.

type NoParam = T0

Useful convenience type for trained models without learnable parameters. Can be used to automatically derive initParam if given as a model's parmaeters.

pattern NoParam :: NoParam