Backprop.Learn.Component.State

Documentation

newtype Unroll :: Nat -> Type -> Type where

Unroll a stateful model into one taking a vector of sequential inputs.

Useful when used before DeState or FixState

Constructors

Unroll :: {..} -> Unroll n l

Fields

getUnroll :: l

newtype DeState :: Type -> Type where

Make a model stateless by converting the state to a trained parameter, and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See FixState, as well.

Constructors

DeState :: {..} -> DeState l

Fields

getDestate :: l

data FixState :: Maybe Type -> Type -> Type where

Make a model stateless by pre-applying a fixed state and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See DeState, as well.

Constructors

FS :: {..} -> FixState s l

Fields

_fsLearn :: l

_fsInitState :: Mayb I s