Backprop.Learn.Component.State

Make models stateless

newtype TrainState :: Type -> Type where

Make a model stateless by converting the state to a trained parameter, and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See DeState, as well.

Its parameters are:

If l has no parameters, just the initial state.

If l has parameters, a T2 of the parameter and initial state.

Constructors

TrainState :: {..} -> TrainState l

Fields

getTrainState :: l

data DeState :: Type -> Type -> Type where

Make a model stateless by pre-applying a fixed state (or a stochastic one with fixed stribution) and dropping the modified state from the result.

One of the ways to make a model stateless for training purposes. Useful when used after Unroll. See TrainState, as well.

Constructors

DS :: {..} -> DeState s l

Fields

_fsInitState :: s

_fsInitStateStoch :: forall m. PrimMonad m => Gen (PrimState m) -> m s

_fsLearn :: l

dsFixed :: s -> l -> DeState s l

Manipulate model states

newtype Unroll :: Nat -> Type -> Type where

Unroll a stateful model into one taking a vector of sequential inputs.

Useful when used before TrainState or DeState. See UnrollTrainState and UnrollDeState.

Constructors

Unroll :: {..} -> Unroll n l

Fields

getUnroll :: l

type UnrollTrainState n l = TrainState (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and treat the initial state as a trained parameter.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollTrainState n l)'

    type LParamMaybe (UnrollDeState n l) = 'TupMaybe (LParamMaybe l) (LStateMaybe l)
    type LStateMaybe (UnrollDeState n l) = 'Nothing

type UnrollDeState n l = DeState (LState l) (Unroll n l)

Unroll a stateful model into a stateless one taking a vector of sequential inputs and fix the initial state.

instance Learn a b l
    => Learn (Vector n a) (Vector n b) (UnrollDeState n l)'

    type LParamMaybe (UnrollDeState n l) = LParamMaybe l
    type LStateMaybe (UnrollDeState n l) = 'Nothing

data ReState :: Type -> Maybe Type -> Type -> Type

Transform the state of a model by providing functions to pre-apply and post-apply before and after the original model sees the state.

I don't really know why you would ever want to do this, but it was fun to write.

Note that a ReState p 'Nothing is essentially the same as DeState p; however, defining one in terms of the other is a bit inconvenient, so they are provided as separate types.

rsFixed :: (forall m. PrimMonad m => Gen (PrimState m) -> Mayb m t) -> (forall q. Reifies q W => BVar q s -> Mayb (BVar q) t) -> (forall q. Reifies q W => Mayb (BVar q) t -> BVar q s) -> l -> ReState s t l

Create a ReState from a fixed (deterministic, non-stochastic) transformation function.