Backprop.Learn.Initialize

Documentation

class Initialize p where

Class for types that are basically a bunch of Doubles, which can be initialized with a given identical and independent distribution.

Methods

initialize :: (ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

initialize :: (ADTRecord p, Constraints p Initialize, ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

gInitialize :: (ADTRecord p, Constraints p Initialize, ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

initialize for any instance of Generic.

initializeNormal

Arguments

:: (Initialize p, PrimMonad m)

=> Double
standard deviation

-> Gen (PrimState m)

-> m p

initializeSingle :: (ContGen d, PrimMonad m, Fractional p) => d -> Gen (PrimState m) -> m p

initialize definition if p is a single number.