Backprop.Learn.Model

Documentation

module Backprop.Learn.Model.Types

module Backprop.Learn.Model.Stochastic

module Backprop.Learn.Model.State

module Backprop.Learn.Model.Regression

module Backprop.Learn.Model.Parameter

module Backprop.Learn.Model.Neural.LSTM

module Backprop.Learn.Model.Neural

module Backprop.Learn.Model.Function

module Backprop.Learn.Model.Combinator

class Backprop a where

Class of values that can be backpropagated in general.

For instances of Num, these methods can be given by zeroNum, addNum, and oneNum. There are also generic options given in Numeric.Backprop.Class for functors, IsList instances, and Generic instances.

instance Backprop Double where
    zero = zeroNum
    add = addNum
    one = oneNum

If you leave the body of an instance declaration blank, GHC Generics will be used to derive instances if the type has a single constructor and each field is an instance of Backprop.

To ensure that backpropagation works in a sound way, should obey the laws:

identity

add x (zero y) = x

add (zero x) y = y

Also implies preservation of information, making zipWith (+) an illegal implementation for lists and vectors.

This is only expected to be true up to potential "extra zeroes" in x and y in the result.

commutativity

add x y = add y x

associativity

add x (add y z) = add (add x y) z

idempotence

zero . zero = zero

one . one = one

Note that not all values in the backpropagation process needs all of these methods: Only the "final result" needs one, for example. These are all grouped under one typeclass for convenience in defining instances, and also to talk about sensible laws. For fine-grained control, use the "explicit" versions of library functions (for example, in Numeric.Backprop.Explicit) instead of Backprop based ones.

This typeclass replaces the reliance on Num of the previous API (v0.1). Num is strictly more powerful than Backprop, and is a stronger constraint on types than is necessary for proper backpropagating. In particular, fromInteger is a problem for many types, preventing useful backpropagation for lists, variable-length vectors (like Data.Vector) and variable-size matrices from linear algebra libraries like hmatrix and accelerate.

Since: backprop-0.2.0.0

Methods

zero :: a -> a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

Should be idempotent:

zero . zero = zero

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See zeroNum for a pre-built definition for instances of Num and zeroFunctor for a definition for instances of Functor. If left blank, will automatically be genericZero, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

add :: a -> a -> a

Add together two values of a type. To combine contributions of gradients, so should be information-preserving:

add x (zero y) = x

add (zero x) y = y

Should be as strict as possible. This behavior is observed for all instances provided by this library.

See addNum for a pre-built definition for instances of Num and addFunctor for a definition for instances of Functor. If left blank, will automatically be genericAdd, a pre-built definition for instances of Generic with one constructor whose fields are all themselves instances of Backprop.

one :: a -> a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

Should be idempotent:

one . one = one

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See oneNum for a pre-built definition for instances of Num and oneFunctor for a definition for instances of Functor. If left blank, will automatically be genericOne, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

Running and Grad

runModel :: forall p s a b. (MaybeC Backprop s, Backprop b) => Model p s a b -> Mayb I p -> a -> Mayb I s -> (b, Mayb I s)

runModelStoch :: forall p s a b m. (MaybeC Backprop s, Backprop b, PrimMonad m) => Model p s a b -> Gen (PrimState m) -> Mayb I p -> a -> Mayb I s -> m (b, Mayb I s)

runModelStateless :: Model p Nothing a b -> Mayb I p -> a -> b

runModelStochStateless :: PrimMonad m => Model p Nothing a b -> Gen (PrimState m) -> Mayb I p -> a -> m b

gradModel :: (Backprop a, Backprop b, MaybeC Backprop p) => Model p Nothing a b -> Mayb I p -> a -> (Mayb I p, a)

gradModelStoch :: (Backprop a, Backprop b, MaybeC Backprop p, PrimMonad m) => Model p Nothing a b -> Gen (PrimState m) -> Mayb I p -> a -> m (Mayb I p, a)

Work with parameters

initParam

Arguments

:: (Initialize p, ContGen d, PrimMonad m)

=> model (Just p) s a b
ignored

-> d

-> Gen (PrimState m)

-> m p

initParamNormal

Arguments

:: (Initialize p, PrimMonad m)

=> model (Just p) s a b
ignored

-> Double

-> Gen (PrimState m)

-> m p

encodeParam

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> p

-> ByteString

decodeParam

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> ByteString

-> p

decodeParamOrFail

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> ByteString

-> Either String p

saveParam

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> FilePath

-> p

-> IO ()

loadParam

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> FilePath

-> IO p

loadParamOrFail

Arguments

:: Binary p

=> model (Just p) s a b
ignored

-> FilePath

-> IO (Either String p)

Iterated runners

iterateModel

Arguments

:: (Backprop b, MaybeC Backprop s)

=> (b -> a)
loop

-> Int
num times

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> ([b], Mayb I s)

iterateModelM

Arguments

:: (Backprop b, MaybeC Backprop s, Monad m)

=> (b -> m a)
loop

-> Int
num times

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> m ([b], Mayb I s)

iterateModelStoch

Arguments

:: (Backprop b, MaybeC Backprop s, PrimMonad m)

=> (b -> m a)
loop

-> Int
num times

-> Model p s a b

-> Gen (PrimState m)

-> Mayb I p

-> a

-> Mayb I s

-> m ([b], Mayb I s)

scanModel :: (Traversable t, Backprop b, MaybeC Backprop s) => Model p s a b -> Mayb I p -> t a -> Mayb I s -> (t b, Mayb I s)

scanModelStoch :: (Traversable t, Backprop b, MaybeC Backprop s, PrimMonad m) => Model p s a b -> Gen (PrimState m) -> Mayb I p -> t a -> Mayb I s -> m (t b, Mayb I s)

No final state

iterateModel_

Arguments

:: (Backprop b, MaybeC Backprop s)

=> (b -> a)
loop

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> [b]

iterateModelM_

Arguments

:: (Backprop b, MaybeC Backprop s, Monad m)

=> (b -> m a)
loop

-> Int
num times

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> m [b]

iterateModelStoch_

Arguments

:: (Backprop b, MaybeC Backprop s, PrimMonad m)

=> (b -> m a)
loop

-> Int
num times

-> Model p s a b

-> Gen (PrimState m)

-> Mayb I p

-> a

-> Mayb I s

-> m [b]

scanModel_ :: (Traversable t, Backprop b, MaybeC Backprop s) => Model p s a b -> Mayb I p -> t a -> Mayb I s -> t b

scanModelStoch_ :: (Traversable t, Backprop b, MaybeC Backprop s, PrimMonad m) => Model p s a b -> Gen (PrimState m) -> Mayb I p -> t a -> Mayb I s -> m (t b)

Prime runners

primeModel :: (Foldable t, Backprop b, MaybeC Backprop s) => Model p s a b -> Mayb I p -> t a -> Mayb I s -> Mayb I s

primeModelStoch :: (Foldable t, Backprop b, MaybeC Backprop s, PrimMonad m) => Model p s a b -> Gen (PrimState m) -> Mayb I p -> t a -> Mayb I s -> m (Mayb I s)

selfPrime

Arguments

:: (Backprop b, MaybeC Backprop s)

=> (b -> a)
loop

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> [Mayb I s]

selfPrimeM

Arguments

:: (Backprop b, MaybeC Backprop s, Monad m)

=> (b -> m a)
loop

-> Int
num times

-> Model p s a b

-> Mayb I p

-> a

-> Mayb I s

-> m (Mayb I s)