Backprop.Learn.Model.Combinator

Documentation

data Feedback :: Type -> Type -> Type -> Type where

Take a model and turn it into a model that runs its output into itself several times, returning the final result. Parameterized by the number of repeats, and the function to process the output to become the next input.

See FeedbackTrace if you want to observe all of the intermediate outputs.

Constructors

FB :: {..} -> Feedback a b l

Fields

_fbTimes :: Int

_fbFeed :: forall s. Reifies s W => BVar s b -> BVar s a

_fbLearn :: l

feedbackId :: Int -> l -> Feedback a a l

Construct a Feedback from an endofunction (a function that returns a value fo the same type as its input) by simply providing the output directly as the next intput.

data FeedbackTrace :: Nat -> Type -> Type -> Type -> Type where

Take a model and turn it into a model that runs its output into itself several times, and returns all of the intermediate outputs. Parameterized by the function to process the output to become the next input.

See Feedback if you only need the final result.

Compare also to Unroll.

Constructors

FBT :: {..} -> FeedbackTrace n a b l

Fields

_fbtFeed :: forall s. Reifies s W => BVar s b -> BVar s a

_fbtLearn :: l

feedbackTraceId :: l -> FeedbackTrace n a a l

Construct a FeedbackTrace from an endofunction (a function that returns a value fo the same type as its input) by simply providing the output directly as the next intput.

data Chain :: [Type] -> [Type] -> [Type] -> Type -> Type -> Type where

Chain components linearly, retaining the ability to deconstruct at a later time.

Constructors

CNil :: Chain '[] '[] '[] a a

(:~>) :: (Learn a b l, KnownMayb (LParamMaybe l), KnownMayb (LStateMaybe l)) => !l -> !(Chain ls ps ss b c) -> Chain (l ': ls) (MaybeToList (LParamMaybe l) ++ ps) (MaybeToList (LStateMaybe l) ++ ss) a c infixr 5

(~++) :: forall ls ms ps qs ss ts a b c. Chain ls ps ss a b -> Chain ms qs ts b c -> Chain (ls ++ ms) (ps ++ qs) (ss ++ ts) a c

Appending Chain

chainParamLength :: Chain ls ps ss a b -> Length ps

chainStateLength :: Chain ls ps ss a b -> Length ss

data LearnFunc :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

Data type representing trainable models.

Useful for performant composition, but you lose the ability to decompose parts.

Constructors

LF :: {..} -> LearnFunc p s a b

Fields

_lfInitParam :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m p

_lfInitState :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m s

_lfRunLearn :: forall q. Reifies q W => Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> (BVar q b, Mayb (BVar q) s)

_lfRunLearnStoch :: forall m q. (PrimMonad m, Reifies q W) => Gen (PrimState m) -> Mayb (BVar q) p -> BVar q a -> Mayb (BVar q) s -> m (BVar q b, Mayb (BVar q) s)

learnFunc :: Learn a b l => l -> LearnFunc (LParamMaybe l) (LStateMaybe l) a b

data LMap :: Type -> Type -> Type -> Type where

Pre-compose a pure parameterless function to a model.

An LMap b a takes a model from b and turns it into a model from a.

Constructors

LM :: {..} -> LMap b a l

Fields

_lmPre :: forall s. Reifies s W => BVar s a -> BVar s b

_lmLearn :: !l

data RMap :: Type -> Type -> Type -> Type where

Post-compose a pure parameterless function to a model.

An Rmap b c takes a model returning b and turns it into a model returning c.

Constructors

RM :: {..} -> RMap b c l

Fields

_rmPost :: forall s. Reifies s W => BVar s b -> BVar s c

_rmLearn :: !l

type Dimap b c a d l = RMap c d (LMap b a l)

Pre- and post-compose pure parameterless functions to a model.

A Dimap b c a d takes a model from b to c and turns it into a model from a to d.

instance Learn b c => Learn a d (Dimap b c a d l) where
    type LParamMaybe (Dimap b c a d l) = LParamMaybe l
    type LStateMaybe (Dimap b c a d l) = LStateMaybe l

pattern DM :: (forall s. Reifies s W => BVar s a -> BVar s b) -> (forall s. Reifies s W => BVar s c -> BVar s d) -> l -> Dimap b c a d l

Constructor for Dimap

_dmPre :: Dimap b c a d l -> forall s. Reifies s W => BVar s a -> BVar s b

_dmPost :: Dimap b c a d l -> forall s. Reifies s W => BVar s c -> BVar s d

_dmLearn :: Dimap b c a d l -> l

(.~) :: forall ps qs ss ts a b c. (ListC (Num <$> ps), ListC (Num <$> qs), ListC (Num <$> ss), ListC (Num <$> ts), ListC (Num <$> (ss ++ ts)), Known Length ps, Known Length qs, Known Length ss, Known Length ts) => LearnFunc (Just (T ps)) (Just (T ss)) b c -> LearnFunc (Just (T qs)) (Just (T ts)) a b -> LearnFunc (Just (T (ps ++ qs))) (Just (T (ss ++ ts))) a c

Compose two LearnFunc on lists.

nilLF :: LearnFunc (Just (T '[])) (Just (T '[])) a a

Identity of .~

onlyLF :: forall p s a b. (KnownMayb p, MaybeC Num p, KnownMayb s, MaybeC Num s) => LearnFunc p s a b -> LearnFunc (Just (T (MaybeToList p))) (Just (T (MaybeToList s))) a b

LearnFunc with singleton lists; meant to be used with .~