Backprop.Learn.Model.Combinator

Tuple-based Composition

(~>) :: forall p q s t a b c. (KnownMayb p, KnownMayb q, KnownMayb s, KnownMayb t, MaybeC Backprop p, MaybeC Backprop q, MaybeC Backprop s, MaybeC Backprop t) => Model p s a b -> Model q t b c -> Model (p :&? q) (s :&? t) a c infixr 8

Compose two Models one after the other, in reverse composition order

(<~) :: forall p q s t a b c. (KnownMayb p, KnownMayb q, KnownMayb s, KnownMayb t, MaybeC Backprop p, MaybeC Backprop q, MaybeC Backprop s, MaybeC Backprop t) => Model p s b c -> Model q t a b -> Model (p :&? q) (s :&? t) a c infixr 8

Compose two Models one after the other.

List-based Composition

type LModel ps ss a b = Model (Just (T ps)) (Just (T ss)) a b

Model where parameters and states are heterogeneous lists (T), making for more seamless composition.

(#:) :: forall p ps s ss a b c. (MaybeC Backprop p, ListC (Backprop <$> ps), MaybeC Backprop s, ListC (Backprop <$> ss), KnownMayb p, KnownMayb s) => Model p s b c -> LModel ps ss a b -> LModel (MaybeToList p ++ ps) (MaybeToList s ++ ss) a c infixr 5

Cons a model to the end of a chain of LModel compositions. Can be used with nilLM.

nilLM :: Model (Just (T '[])) (Just (T '[])) a a

Identity of #++

(#++) :: forall ps qs ss ts a b c. (ListC (Backprop <$> ps), ListC (Backprop <$> qs), ListC (Backprop <$> ss), ListC (Backprop <$> ts), ListC (Backprop <$> (ps ++ qs)), ListC (Backprop <$> (ss ++ ts)), Known Length ps, Known Length ss, Known Length ts) => LModel ps ss b c -> LModel qs ts a b -> LModel (ps ++ qs) (ss ++ ts) a c infixr 5

Compose two LModels

liftLM :: forall p s a b. (KnownMayb p, MaybeC Backprop p, KnownMayb s, MaybeC Backprop s) => Model p s a b -> LModel (MaybeToList p) (MaybeToList s) a b

Lift a normal Model to a LModel with a singleton list parameter/state if 'Just, or an empty list if 'Nothing. Essentially prepares a model to be used with #~ and nilLM.

Misc

forkModel

Arguments

:: (KnownMayb p, KnownMayb q, KnownMayb s, KnownMayb t, MaybeC Backprop p, MaybeC Backprop q, MaybeC Backprop s, MaybeC Backprop t, Backprop b, Backprop c)

=> Model p s a b
fork 1

-> Model q t a c
fork 2

-> Model (p :&? q) (s :&? t) a (b :& c)

feedback

Arguments

:: (KnownMayb p, KnownMayb q, KnownMayb s, KnownMayb t, MaybeC Backprop p, MaybeC Backprop q, MaybeC Backprop s, MaybeC Backprop t)

=> Int
times

-> Model p s a b
feed

-> Model q t b a
back

-> Model (p :&? q) (s :&? t) a b

Return a model that loops a model ("feed") repeatedly onto itself, with a model to provide the back loop.

feedbackTrace

Arguments

:: (KnownMayb p, KnownMayb q, KnownMayb s, KnownMayb t, MaybeC Backprop p, MaybeC Backprop q, MaybeC Backprop s, MaybeC Backprop t, KnownNat n, Backprop b)

=> Model p s a b
feed

-> Model q t b a
back

-> Model (p :&? q) (s :&? t) a (ABP (Vector n) b)

feedback, but tracing and observing all of the intermediate values.