Backprop.Learn.Model.Neural

Fully connected

Feed-forward

type FC i o = LinReg i o

Fully connected feed-forward layer with bias. Parameterized by its initialization distribution.

Note that this has no activation function; to use as a model with activation function, chain it with an activation function using RMap, :.~, etc.; see FCA for a convenient type synonym and constructor.

Without any activation function, this is essentially a multivariate linear regression.

With the logistic function as an activation function, this is essentially multivariate logistic regression. (See Logistic)

pattern FC :: FC i o

type FCp = LRp

fcBias :: Lens' (FCp i o) (R o)

fcWeights :: Lens (FCp i o) (FCp i' o) (L o i) (L o i')

With activation function

type FCA i o = RMap (R o) (R o) (FC i o)

Convenient synonym for an FC post-composed with a simple parameterless activation function.

pattern FCA :: (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCA i o

Construct an FCA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

_fcaActivation :: FCA i o -> forall s. Reifies s W => BVar s (R o) -> BVar s (R o)

Recurrent

type FCR h i o = Recurrent (R (i + h)) (R i) (R h) (R o) (FC (i + h) o)

Fully connected recurrent layer with bias.

Parameterized by its initialization distributions, and also the function to compute the new state from previous input.

instance Learn (R i) (R o) (FCR h i o) where
    type LParamMaybe (FCR h i o) = 'Just (FCRp h i o)
    type LStateMaybe (FCR h i o) = 'Just (R h)

pattern FCR :: (KnownNat h, KnownNat i) => (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> FCR h i o

Construct an FCR

type FCRp h i o = FCp (i + h) o

fcrBias :: Lens' (FCRp h i o) (R o)

fcrInputWeights :: (KnownNat h, KnownNat i, KnownNat i', KnownNat o) => Lens (FCRp h i o) (FCRp h i' o) (L o i) (L o i')

fcrStateWeights :: (KnownNat h, KnownNat h', KnownNat i, KnownNat o) => Lens (FCRp h i o) (FCRp h' i o) (L o h) (L o h')

With activation function

type FCRA h i o = RMap (R o) (R o) (FCR h i o)

Convenient synonym for an FCR post-composed with a simple parameterless activation function.

pattern FCRA :: (KnownNat h, KnownNat i) => (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCRA h i o

Construct an FCRA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

_fcraStore :: FCRA h i o -> forall s. (Reifies s W, KnownNat h, KnownNat i) => BVar s (R o) -> BVar s (R h)

_fcraActivation :: FCRA h i o -> forall s. (Reifies s W, KnownNat h, KnownNat i) => BVar s (R o) -> BVar s (R o)