Backprop.Learn.Model.Neural

Fully connected

Feed-forward

newtype FC (i :: Nat) (o :: Nat)

Fully connected feed-forward layer with bias. Parameterized by its initialization distribution.

Note that this has no activation function; to use as a model with activation function, chain it with an activation function using RMap, :.~, etc.; see FCA for a convenient type synonym and constructor.

Without any activation function, this is essentially a multivariate linear regression.

With the logistic function as an activation function, this is essentially multivariate logistic regression. (See Logistic)

Constructors

FC

Fields

_fcGen :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

fc :: ContGen d => d -> FC i o

Construct an FC i o using a given distribution from the statistics library.

data FCp i o

Fully connected feed-forward layer parameters.

Constructors

FCp

Fields

_fcBias :: !(R o)

_fcWeights :: !(L o i)

fcBias :: Lens' (FCp i o) (R o)

fcWeights :: Lens (FCp i o) (FCp i' o) (L o i) (L o i')

With activation function

type FCA i o = RMap (R o) (R o) (FC i o)

Convenient synonym for an FC post-composed with a simple parameterless activation function.

pattern FCA :: (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCA i o

Construct an FCA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

fca :: ContGen d => d -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCA i o

Construct an FCA i o using a given distribution from the statistics library.

_fcaGen :: FCA i o -> forall (m :: * -> *). PrimMonad m => Gen (PrimState m) -> m Double

_fcaActivation :: FCA i o -> forall s. Reifies s W => BVar s (R o) -> BVar s (R o)

Recurrent

data FCR (h :: Nat) (i :: Nat) (o :: Nat)

Fully connected recurrent layer with bias.

Parameterized by its initialization distributions, and also the function to compute the new state from previous input.

Constructors

FCR

Fields

_fcrGen :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_fcrGenState :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_fcrStore :: forall s. Reifies s W => BVar s (R o) -> BVar s (R h)

fcr :: (ContGen d, ContGen e) => d -> e -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> FCR h i o

Construct an FCR h i o using given distributions from the statistics library.

data FCRp h i o

Fully connected recurrent layer parameters.

Constructors

FCRp

Fields

_fcrBias :: !(R o)

_fcrInputWeights :: !(L o i)

_fcrStateWeights :: !(L o h)

fcrBias :: Lens' (FCRp h i o) (R o)

fcrInputWeights :: Lens (FCRp h i o) (FCRp h i' o) (L o i) (L o i')

fcrStateWeights :: Lens (FCRp h i o) (FCRp h' i o) (L o h) (L o h')

With activation function

type FCRA h i o = RMap (R o) (R o) (FCR h i o)

Convenient synonym for an FCR post-composed with a simple parameterless activation function.

pattern FCRA :: (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCRA h i o

Construct an FCRA using a generating function and activation function.

Some common ones include logistic and vmap reLU.

fcra :: (ContGen d, ContGen e) => d -> e -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R o)) -> FCRA h i o

Construct an FCRA h i o using given distributions from the statistics library.

_fcraGen :: FCRA h i o -> forall (m :: * -> *). PrimMonad m => Gen (PrimState m) -> m Double

_fcraGenState :: FCRA h i o -> forall (m :: * -> *). PrimMonad m => Gen (PrimState m) -> m Double

_fcraStore :: FCRA h i o -> forall s. Reifies s W => BVar s (R o) -> BVar s (R h)

_fcraActivation :: FCRA h i o -> forall s. Reifies s W => BVar s (R o) -> BVar s (R o)