Backprop.Learn.Model.Neural

Documentation

newtype FC (i :: Nat) (o :: Nat)

Fully connected feed-forward layer with bias. Parameterized by its initialization distribution.

Note that this has no activation function; to use as a model with activation function, chain it with an activation function using Dimap, :.~, etc.

Without any activation function, this is essentially a multivariate linear regression.

With the logistic function as an activation function, this is essentially multivariate logistic regression. (See Logistic)

Constructors

FC

Fields

_fcGen :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

fc :: ContGen d => d -> FC i o

Construct an FC i o using a given distribution from the statistics library.

data FCp i o

Fully connected feed-forward layer parameters.

Constructors

FCp

Fields

_fcBias :: !(R o)

_fcWeights :: !(L o i)

fcBias :: Functor f => (R o -> f (R o)) -> FCp i o -> f (FCp i o)

fcWeights :: Functor f => (L o i -> f (L o k)) -> FCp i o -> f (FCp k o)

data FCR (h :: Nat) (i :: Nat) (o :: Nat)

Fully connected recurrent layer with bias.

Parameterized by its initialization distributions, and also the function to compute the new state from previous input.

Constructors

FCR

Fields

_fcrGen :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_fcrGenState :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_fcrStore :: forall s. Reifies s W => BVar s (R o) -> BVar s (R h)

fcr :: ContGen d => d -> (forall s. Reifies s W => BVar s (R o) -> BVar s (R h)) -> FCR h i o

Construct an FCR h i o using a given distribution from the statistics library.

data FCRp h i o

Fully connected recurrent layer parameters.

Constructors

FCRp

Fields

_fcrBias :: !(R o)

_fcrInputWeights :: !(L o i)

_fcrStateWeights :: !(L o h)

fcrBias :: Functor f => (R o -> f (R o)) -> FCRp h i o -> f (FCRp h i o)

fcrInputWeights :: Functor f => (L o i -> f (L o i')) -> FCRp h i o -> f (FCRp h i' o)

fcrStateWeights :: Functor f => (L o h -> f (L o h')) -> FCRp h i o -> f (FCRp h' i o)