Backprop.Learn.Model.Neural.LSTM

Documentation

type LSTM i o = Recurrent (R (i + o)) (R i) (R o) (R o) (LSTM' i o)

Long-term short-term memory layer

http://colah.github.io/posts/2015-08-Understanding-LSTMs/

instance Learn (R i) (R o) (LSTM i o) where
    type LParamMaybe (LSTM i o) = 'Just (LSTMp i o)
    type LStateMaybe (LSTM i o) = 'Just (T2 (R o) (R o))     -- cell state, hidden state

pattern LSTM :: (KnownNat i, KnownNat o) => (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> (forall m. PrimMonad m => Gen (PrimState m) -> m (R o)) -> Bool -> LSTM i o

Construct an LSTM

lstm

Arguments

:: (KnownNat i, KnownNat o, ContGen d, ContGen e, ContGen f)

=> d
generate parameters

-> e
generate initial cell state

-> f
generate initial history

-> Bool
forgiet bias initialized to 1

-> LSTM i o

Convenient wrapper over LSTM constructor taking distributions from the statistics library.

_lstmGen :: LSTM i o -> forall (m :: * -> *). (PrimMonad m, KnownNat i, KnownNat o) => Gen (PrimState m) -> m Double

_lstmGenCell :: LSTM i o -> forall (m :: * -> *). (PrimMonad m, KnownNat i, KnownNat o) => Gen (PrimState m) -> m Double

_lstmGenState :: LSTM i o -> forall (m :: * -> *). (PrimMonad m, KnownNat i, KnownNat o) => Gen (PrimState m) -> m (R o)

_lstmForgetBias :: LSTM i o -> KnownNat i -> KnownNat o -> Bool

data LSTMp (i :: Nat) (o :: Nat)

LSTM layer parmateters

Constructors

LSTMp

Fields

_lstmForget :: !(FCp (i + o) o)

_lstmInput :: !(FCp (i + o) o)

_lstmUpdate :: !(FCp (i + o) o)

_lstmOutput :: !(FCp (i + o) o)

lstmForget :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmInput :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmUpdate :: Lens' (LSTMp i o) (FCp (i + o) o)

lstmOutput :: Lens' (LSTMp i o) (FCp (i + o) o)

data LSTM' (i :: Nat) (o :: Nat)

Base for LSTM. An LSTM is just an LSTM' where the second half of the input vector is the previous output.

This is mostly an implementation detail. It is recommended that you use LSTM and its constructor, instead of directly using this.

Constructors

LSTM'

Fields

_lstmGen' :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_lstmGenCell' :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_lstmForgetBias' :: Bool