Backprop.Learn.Model.Parameter

Documentation

data DeParam :: Type -> Type -> Type where

Convert a model with trainabile parameters into a model without any trainable parameters.

The parameters are instead fixed (or stochastic, with a fixed distribution), and the model is treated as an untrainable function.

DeParam is essentially DeParamAt, with id as the lens.

Constructors

DP :: {..} -> DeParam p l

Fields

_dpParam :: p

_dpParamStoch :: forall m. PrimMonad m => Gen (PrimState m) -> m p

_dpLearn :: l

dpDeterm :: p -> l -> DeParam p l

Create a DeParam from a deterministic, non-stochastic parameter.

data DeParamAt :: Type -> Type -> Type -> Type -> Type where

Wrapping a mode with DeParamAt pq p q says that the mode's parameter pq can be split into p and q, and fixes q to a given fixed (or stochastic with a fixed distribution) value. The model now effectively has parameter p only, and the q part will not be backpropagated.

DeParam is essentially DeParamAt where p is '()' or T0.

Constructors

DPA :: {..} -> DeParamAt pq p q l

Fields

_dpaSplit :: pq -> (p, q)

_dpaJoin :: p -> q -> pq

_dpaParam :: q

_dpaParamStoch :: forall m. PrimMonad m => Gen (PrimState m) -> m q

_dpaLearn :: l

dpaDeterm :: (pq -> (p, q)) -> (p -> q -> pq) -> q -> l -> DeParamAt pq p q l

Create a DeParamAt from a deterministic, non-stochastic fixed value as a part of the parameter.

data ReParam :: Type -> Maybe Type -> Type -> Type

Pre-apply a function to the parameter before the original model sees it. A ReParam p q turns a model taking p into a model taking q.

Note that a ReParam p 'Nothing is essentially the same as a DeParam p, and one could implement DeParamAt p q in terms of ReParam p ('Just q).

rpDeterm :: (forall s. Reifies s W => Mayb (BVar s) q -> BVar s p) -> l -> ReParam p q l

Create a ReParam from a deterministic, non-stochastic transformation function.