Backprop.Learn.Model.Regression

Documentation

type Linear n m = FC n m

Multivariate linear regression, from an n-vector to an m-vector.

Is essentially just a fully connected neural network layer with bias, with no activation function.

linearRegression :: (forall f. PrimMonad f => Gen (PrimState f) -> f Double) -> Linear n m

Construct a linear regression model from an initialization function for coefficients

type Logistic n = RMap (R 1) Double (Linear n 1)

logisticRegression :: (forall m. PrimMonad m => Gen (PrimState m) -> m Double) -> Logistic n

Construct a logistic regression model from an initialization function for coefficients.

data ARMA :: Nat -> Nat -> Type where

Auto-regressive moving average model.

ARMA(p,q) is an ARMA model with p autoregressive history terms and q error (innovation) history terms.

It is a Learn Double Double instance, and is meant to predict the "next step" of a time sequence.

In this state, it is a runnable stateful model. To train, use with ARMAUnroll to unroll and fix the initial state.

An ARIMA (auto-regressive moving average model on differenced history) model is a small modification from this; it is theoretically possible, I just haven't gotten around to implementing it yet! Once done, we could just define the type synonym type ARMA p q = ARIMA p 0 q.

Constructors

ARMA :: {..} -> ARMA p q

Fields

_armaGenPhi :: forall m. PrimMonad m => Gen (PrimState m) -> Finite p -> m Double

_armaGenTheta :: forall m. PrimMonad m => Gen (PrimState m) -> Finite q -> m Double

_armaGenConst :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_armaGenYHist :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_armaGenEHist :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

data ARMAp :: Nat -> Nat -> Type where

ARMA parmaeters

Constructors

ARMAp :: {..} -> ARMAp p q

Fields

_armaPhi :: !(R p)

_armaTheta :: !(R q)

_armaConstant :: !Double

data ARMAs :: Nat -> Nat -> Type where

ARMA state

Constructors

ARMAs :: {..} -> ARMAs p q

Fields

_armaYPred :: !Double

_armaYHist :: !(R p)

_armaEHist :: !(R q)

type ARMAUnroll p q = DeParamAt (T2 (ARMAp p q) (ARMAs p q)) (R q) (UnrollTrainState (Max p q) (ARMA p q))

The "unrolled" and "destated" ARMA model, which takes a vector of sequential inputs and outputs a vector of the model's expected next steps.

Useful for actually training ARMA using gradient descent.

This fixes the initial error history to be zero (or a fixed stochastic sample), and treats the initial output history to be a /learned parameter/.

instance Learn (Vector n Double) (SV.Vector n Double) (ARMAUnroll p q) where
    -- | Initial state is a parameter, but initial error history is fixed
    type LParamMaybe (ARMAUnroll p q) = 'Just (T2 (ARMAp p q) (ARMAs p q))
    type LStateMaybe (ARMAUnroll p q) = 'Nothing

armaUnroll :: KnownNat q => ARMA p q -> ARMAUnroll p q

Constructor for ARMAUnroll

type ARMAUnrollFinal p q = DeParamAt (T2 (ARMAp p q) (ARMAs p q)) (R q) (UnrollFinalTrainState (Max p q) (ARMA p q))

ARMAUnroll, but only looking at the final output after running the model on all inputs.

instance Learn (Vector n Double) Double (ARMAUnrollFinal p q) where
    -- | Initial state is a parameter, but initial error history is fixed
    type LParamMaybe (ARMAUnrollFinal p q) = 'Just (T2 (ARMAp p q) (ARMAs p q))
    type LStateMaybe (ARMAUnrollFinal p q) = 'Nothing

armaUnrollFinal :: KnownNat q => ARMA p q -> ARMAUnrollFinal p q

Constructor for ARMAUnrollFinal

type AR p = ARMA p 0

Autoregressive model

type MA q = ARMA 0 q

Moving average model