Backprop.Learn.Model.Regression

Documentation

data LinReg (i :: Nat) (o :: Nat)

Multivariate linear regression, from an i-vector to an o-vector.

Constructors

LinReg

type LogReg i o = RMap (R o) (R o) (LinReg i o)

Mutivariate Logistic regression, from an i-vector to an o-vector.

Essentially a linear regression postcomposed with the logistic function.

pattern LogReg :: LogReg i o

Constructor for a LogReg

data LRp i o

Linear Regression parameter

Constructors

LRp

Fields

_lrAlpha :: !(R o)

_lrBeta :: !(L o i)

lrBeta :: Lens (LRp i o) (LRp i' o) (L o i) (L o i')

lrAlpha :: Lens' (LRp i o) (R o)

runLRp :: (KnownNat i, KnownNat o, Reifies s W) => BVar s (LRp i o) -> BVar s (R i) -> BVar s (R o)

data ARIMA :: Nat -> Nat -> Nat -> Type where

Auto-regressive integrated moving average model.

ARIMA(p,d,q) is an ARIMA model with p autoregressive history terms on the d-th order differenced history and q error (innovation) history terms.

It is a Learn Double Double instance, and is meant to predict the "next step" of a time sequence.

In this state, it is a runnable stateful model. To train, use with ARIMAUnroll to unroll and fix the initial state.

Constructors

ARIMA :: {..} -> ARIMA p d q

Fields

_arimaGenYHist :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

_arimaGenEHist :: forall m. PrimMonad m => Gen (PrimState m) -> m Double

data ARIMAp :: Nat -> Nat -> Type where

ARIMA parmaeters

Constructors

ARIMAp :: {..} -> ARIMAp p q

Fields

_arimaPhi :: !(R p)

_arimaTheta :: !(R q)

_arimaConstant :: !Double

data ARIMAs :: Nat -> Nat -> Nat -> Type where

ARIMA state

Constructors

ARIMAs :: {..} -> ARIMAs p d q

Fields

_arimaYPred :: !Double

_arimaYHist :: !(R (p + d))

_arimaEHist :: !(R q)

type ARIMAUnroll p d q = DeParamAt (T2 (ARIMAp p q) (ARIMAs p d q)) (R q) (UnrollTrainState (Max (p + d) q) (ARIMA p d q))

The "unrolled" and "destated" ARIMA model, which takes a vector of sequential inputs and outputs a vector of the model's expected next steps.

Useful for actually training ARIMA using gradient descent.

This fixes the initial error history to be zero (or a fixed stochastic sample), and treats the initial output history to be a /learned parameter/.

instance Learn (Vector n Double) (SV.Vector n Double) (ARIMAUnroll p q) where
    -- | Initial state is a parameter, but initial error history is fixed
    type LParamMaybe (ARIMAUnroll p q) = 'Just (T2 (ARIMAp p q) (ARIMAs p q))
    type LStateMaybe (ARIMAUnroll p q) = 'Nothing

arimaUnroll :: KnownNat q => ARIMA p d q -> ARIMAUnroll p d q

Constructor for ARIMAUnroll

type ARIMAUnrollFinal p d q = DeParamAt (T2 (ARIMAp p q) (ARIMAs p d q)) (R q) (UnrollFinalTrainState (Max (p + d) q) (ARIMA p d q))

ARIMAUnroll, but only looking at the final output after running the model on all inputs.

instance Learn (Vector n Double) Double (ARIMAUnrollFinal p q) where
    -- | Initial state is a parameter, but initial error history is fixed
    type LParamMaybe (ARIMAUnrollFinal p q) = 'Just (T2 (ARIMAp p q) (ARIMAs p q))
    type LStateMaybe (ARIMAUnrollFinal p q) = 'Nothing

arimaUnrollFinal :: KnownNat q => ARIMA p d q -> ARIMAUnrollFinal p d q

Constructor for ARIMAUnrollFinal

type AR p = ARIMA p 0 0

Autoregressive model

type MA = ARIMA 0 0

Moving average model

type ARMA p = ARIMA p 0

Autoregressive Moving average model