Backprop.Learn.Model.Stochastic

Documentation

newtype DO (n :: Nat)

Dropout layer. Parameterized by dropout percentage (should be between 0 and 1).

0 corresponds to no dropout, 1 corresponds to complete dropout of all nodes every time.

Constructors

DO

Fields

_doRate :: Double

data StochFunc :: Maybe Type -> Type -> Type -> Type where

Represents a random-valued function, with a possible trainable parameter.

Requires both a "deterministic" and a "stochastic" mode. The deterministic mode ideally should approximate some mean of the stochastic mode.

Constructors

SF :: {..} -> StochFunc p a b

Fields

_sfInitParam :: forall m. PrimMonad m => Gen (PrimState m) -> Mayb m p

_sfRunDeterm :: forall s. Reifies s W => Mayb (BVar s) p -> BVar s a -> BVar s b

_sfRunStoch :: forall m s. (PrimMonad m, Reifies s W) => Gen (PrimState m) -> Mayb (BVar s) p -> BVar s a -> m (BVar s b)

type FixedStochFunc = StochFunc Nothing

Convenient alias for a StochFunc (random-valued function with both deterministic and stochastic modes) with no trained parameters.

pattern FSF :: (forall s. Reifies s W => BVar s a -> BVar s b) -> (forall m s. (PrimMonad m, Reifies s W) => Gen (PrimState m) -> BVar s a -> m (BVar s b)) -> FixedStochFunc a b

Construct a FixedStochFunc

_fsfRunDeterm :: FixedStochFunc a b -> forall s. Reifies s W => BVar s a -> BVar s b

_fsfRunStoch :: FixedStochFunc a b -> forall (m :: * -> *) s. (PrimMonad m, Reifies s W) => Gen (PrimState m) -> BVar s a -> m (BVar s b)

rreLU :: (ContGen d, Mean d, KnownNat n) => d -> FixedStochFunc (R n) (R n)

Random leaky rectified linear unit

injectNoise :: (ContGen d, Mean d, KnownNat n) => d -> FixedStochFunc (R n) (R n)

Inject random noise. Usually used between neural network layers, or at the very beginning to pre-process input.

In non-stochastic mode, this adds the mean of the distribution.

applyNoise :: (ContGen d, Mean d, KnownNat n) => d -> FixedStochFunc (R n) (R n)

Multply by random noise. Can be used to implement dropout-like behavior.

In non-stochastic mode, this scales by the mean of the distribution.