Backprop.Learn.Model.Types

Model type

type ModelFunc p s a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> (BVar z b, PMaybe (BVar z) s)

type ModelFuncStoch p s a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> m (BVar z b, PMaybe (BVar z) s)

data Model :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

General parameterized model with potential state

Constructors

Model

Fields

:: { runLearn :: ModelFunc p s a b

, runLearnStoch :: ModelFuncStoch p s a b

} -> Model p s a b

modelD :: ModelFunc p s a b -> Model p s a b

Construct a deterministic model, with no stochastic component.

Specialized Models

Stateless

type ModelFuncStateless p a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> BVar z b

type ModelFuncStochStateless p a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> m (BVar z b)

type ModelStateless p = Model p Nothing

Parameterized model with no state

pattern ModelStateless :: ModelFuncStateless p a b -> ModelFuncStochStateless p a b -> ModelStateless p a b

Construct a ModelStateless.

runLearnStateless :: ModelStateless p a b -> forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> BVar z b

runLearnStochStateless :: ModelStateless p a b -> forall (m :: Type -> Type) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> PMaybe (BVar z) p -> BVar z a -> m (BVar z b)

modelStatelessD :: ModelFuncStateless p a b -> ModelStateless p a b

Construct a deterministic stateless model, with no stochastic component.

Stateless and Parameterless

type BFunc a b = forall z. Reifies z W => BVar z a -> BVar z b

type BFuncStoch a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

type Func = Model Nothing Nothing

pattern Func :: BFunc a b -> BFuncStoch a b -> Func a b

Unparameterized model with no state

runFunc :: Func a b -> forall z. Reifies z W => BVar z a -> BVar z b

runFuncStoch :: Func a b -> forall (m :: Type -> Type) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

funcD :: BFunc a b -> Func a b

Construct an deterministic unparameterized stateless model, with no stochastic component.

Manipulating models as functions

type ModelFuncM m p s a b = forall z. Reifies z W => PMaybe (BVar z) p -> BVar z a -> PMaybe (BVar z) s -> m (BVar z b, PMaybe (BVar z) s)

withModelFunc0 :: (forall m. Monad m => ModelFuncM m p s a b) -> Model p s a b

withModelFunc :: (forall m. Monad m => ModelFuncM m p s a b -> ModelFuncM m q t c d) -> Model p s a b -> Model q t c d

withModelFunc2 :: (forall m. Monad m => ModelFuncM m p s a b -> ModelFuncM m q t c d -> ModelFuncM m r u e f) -> Model p s a b -> Model q t c d -> Model r u e f

Utility

data PMaybe (a :: k -> Type) (b :: Maybe k) :: forall k. (k -> Type) -> Maybe k -> Type where

A PMaybe f 'Nothing contains nothing, and a PMaybe f ('Just a) contains an f a.

In practice this can be useful to write polymorphic functions/abstractions that contain an argument that can be "turned off" for different instances.

Constructors

PNothing :: forall k (a :: k -> Type) (b :: Maybe k). PMaybe a (Nothing :: Maybe k)

PJust :: forall k (a :: k -> Type) (b :: Maybe k) (a1 :: k). a a1 -> PMaybe a (Just a1)

Bundled Patterns

pattern TJust :: a -> TMaybe (Just a)

type TMaybe = PMaybe TF

fromPJust :: PMaybe f (Just a) -> f a

type Learnables as = (RecApplicative as, ReifyConstraint Backprop TF as, RMap as, RApply as)

Combination of common constraints for type-level lists.

class (Initialize a, Regularize a, Binary a, NFData a, LinearInPlace m Double a) => Learnable m a

Helpful utility class for ensuring that a parameter or state type is Learnable

class Backprop p => Regularize p where

A class for data types that support regularization during training.

This class is somewhat similar to Metric Double, in that it supports summing the components and summing squared components. However, the main difference is that when summing components, we only consider components that we want to regularize.

Often, this doesn't include bias terms (terms that "add" to inputs), and only includes terms that "scale" inputs, like components in a weight matrix of a feed-forward neural network layer.

However, if all of your components are to be regularized, you can use norm_1, norm_2, lassoLinear, and ridgeLinear as sensible implementations, or use DerivingVia with RegularizeMetric:

data MyType = ...
  deriving Regularize via (RegularizeMetric MyType)

You can also derive an instance where no are regularized, using NoRegularize:

data MyType = ...
  deriving Regularize via (NoRegularize MyType)

The default implementations are based on Generics, and work for types that are records of items that are all instances of Regularize.

Minimal complete definition

Nothing

Methods

rnorm_1 :: p -> Double

Like norm_1: sums all of the weights in p, but only the ones you want to regularize:

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_1 = norm_1. However, this would count all terms in p, even potential bias terms.

rnorm_1 :: (ADT p, Constraints p Regularize) => p -> Double

Like norm_1: sums all of the weights in p, but only the ones you want to regularize:

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_1 = norm_1. However, this would count all terms in p, even potential bias terms.

rnorm_2 :: p -> Double

Like norm_2: sums all of the squares of the weights n p, but only the ones you want to regularize:

\[ \sum_w w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_2 = norm_2. However, this would count all terms in p, even potential bias terms.

rnorm_2 :: (ADT p, Constraints p Regularize) => p -> Double

Like norm_2: sums all of the squares of the weights n p, but only the ones you want to regularize:

\[ \sum_w w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

If p is an instance of Metric, then you can set rnorm_2 = norm_2. However, this would count all terms in p, even potential bias terms.

lasso :: Double -> p -> p

lasso r p sets all regularized components (that is, components summed by rnorm_1) in p to be either r if that component was positive, or -r if that component was negative. Behavior is not defined if the component is exactly zero, but either r or -r are sensible possibilities.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_1 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set lasso = lassoLinear. However, this is only valid if rnorm_1 counts all terms in p, including potential bias terms.

lasso :: (ADT p, Constraints p Regularize) => Double -> p -> p

lasso r p sets all regularized components (that is, components summed by rnorm_1) in p to be either r if that component was positive, or -r if that component was negative. Behavior is not defined if the component is exactly zero, but either r or -r are sensible possibilities.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_1 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set lasso = lassoLinear. However, this is only valid if rnorm_1 counts all terms in p, including potential bias terms.

ridge :: Double -> p -> p

ridge r p scales all regularized components (that is, components summed by rnorm_2) in p by r.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_2 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set ridge = ridgeLinear. However, this is only valid if rnorm_2 counts all terms in p, including potential bias terms.

ridge :: (ADT p, Constraints p Regularize) => Double -> p -> p

ridge r p scales all regularized components (that is, components summed by rnorm_2) in p by r.

It must set all non-regularized components (like bias terms, or whatever items that rnorm_2 ignores) to zero.

If p is an instance of Linear Double and Num, then you can set ridge = ridgeLinear. However, this is only valid if rnorm_2 counts all terms in p, including potential bias terms.

class Initialize p where

Class for types that are basically a bunch of Doubles, which can be initialized with a given identical and independent distribution.

Minimal complete definition

Nothing

Methods

initialize :: (ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p

initialize :: (ADTRecord p, Constraints p Initialize, ContGen d, PrimMonad m) => d -> Gen (PrimState m) -> m p