Backprop.Learn.Model.Types

Model type

type ModelFunc p s a b = forall z. Reifies z W => Mayb (BVar z) p -> BVar z a -> Mayb (BVar z) s -> (BVar z b, Mayb (BVar z) s)

type ModelFuncStoch p s a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> Mayb (BVar z) p -> BVar z a -> Mayb (BVar z) s -> m (BVar z b, Mayb (BVar z) s)

data Model :: Maybe Type -> Maybe Type -> Type -> Type -> Type where

General parameterized model with potential state

Constructors

Model :: {..} -> Model p s a b

Fields

runLearn :: ModelFunc p s a b

runLearnStoch :: ModelFuncStoch p s a b

modelD :: ModelFunc p s a b -> Model p s a b

Construct a deterministic model, with no stochastic component.

Specialized Models

Stateless

type ModelFuncStateless p a b = forall z. Reifies z W => Mayb (BVar z) p -> BVar z a -> BVar z b

type ModelFuncStochStateless p a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> Mayb (BVar z) p -> BVar z a -> m (BVar z b)

type ModelStateless p = Model p Nothing

Parameterized model with no state

pattern ModelStateless :: ModelFuncStateless p a b -> ModelFuncStochStateless p a b -> ModelStateless p a b

Construct a ModelStateless.

runLearnStateless :: ModelStateless p a b -> forall z. Reifies z W => Mayb (BVar z) p -> BVar z a -> BVar z b

runLearnStochStateless :: ModelStateless p a b -> forall (m :: * -> *) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> Mayb (BVar z) p -> BVar z a -> m (BVar z b)

modelStatelessD :: ModelFuncStateless p a b -> ModelStateless p a b

Construct a deterministic stateless model, with no stochastic component.

Stateless and Parameterless

type BFunc a b = forall z. Reifies z W => BVar z a -> BVar z b

type BFuncStoch a b = forall m z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

type Func = Model Nothing Nothing

pattern Func :: BFunc a b -> BFuncStoch a b -> Func a b

Unparameterized model with no state

runFunc :: Func a b -> forall z. Reifies z W => BVar z a -> BVar z b

runFuncStoch :: Func a b -> forall (m :: * -> *) z. (PrimMonad m, Reifies z W) => Gen (PrimState m) -> BVar z a -> m (BVar z b)

funcD :: BFunc a b -> Func a b

Construct an deterministic unparameterized stateless model, with no stochastic component.

Utility

data Mayb :: (k -> Type) -> Maybe k -> Type where

Constructors

N_ :: Mayb f Nothing

J_ :: !(f a) -> Mayb f (Just a)

Bundled Patterns

pattern J_I :: a -> Mayb I (Just a)

fromJ_ :: Mayb f (Just a) -> f a

type family MaybeC (c :: k -> Constraint) (m :: Maybe k) :: Constraint where ...

Equations

MaybeC c (Just a) = c a

MaybeC c Nothing = ()

type KnownMayb = Known (Mayb P)

knownMayb :: KnownMayb p => Mayb P p

newtype I a

Constructors

I

Fields

getI :: a

type family HKD f a where ...

Helper type family for HKD data types. See http://reasonablypolymorphic.com/blog/higher-kinded-data

Equations

HKD I a = a

HKD f a = f a