Backprop.Learn.Regularize

Documentation

type Regularizer p = forall s. Reifies s W => BVar s p -> BVar s Double

A regularizer on parameters

class Backprop p => Regularize p where

Minimal complete definition

l2Reg

Methods

l1Reg :: Double -> Regularizer p

L1 regularization

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

l1Reg :: (ADT p, Constraints p Regularize) => Double -> Regularizer p

L1 regularization

\[ \sum_w \lvert w \rvert \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

l2Reg :: Double -> Regularizer p

L2 regularization

\[ \sum_w \frac{1}{2} w^2 \]

Note that typically bias terms (terms that add to inputs) are not regularized. Only "weight" terms that scale inputs are typically regularized.

noReg :: Regularizer p

No regularization

l2RegDefault

Arguments

:: (Metric Double p, Backprop p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

Defalt L2 regularization for instances of Metric. This will count all terms, including any potential bias terms.

l1RegDefault

Arguments

:: (Num p, Metric Double p, Backprop p)

=> Double
scaling factor (often 0.5)

-> Regularizer p

Defalt L1 regularization for instances of Metric. This will count all terms, including any potential bias terms.

Manipulate regularizers

addReg :: Regularizer p -> Regularizer p -> Regularizer p

Add together two regularizers

scaleReg :: Double -> Regularizer p -> Regularizer p

Scale a regularizer's influence