Backprop.Learn.Train

Gradients

gradLearnLoss :: (Learn a b l, NoState l, LParamMaybe l ~ Just (LParam l), Backprop (LParam l)) => Loss b -> Regularizer (LParam l) -> l -> LParam l -> a -> b -> LParam l

Gradient of model with respect to loss function and target

gradLearnStochLoss :: (Learn a b l, NoState l, LParamMaybe l ~ Just (LParam l), Backprop (LParam l), PrimMonad m) => Loss b -> Regularizer (LParam l) -> l -> Gen (PrimState m) -> LParam l -> a -> b -> m (LParam l)

Stochastic gradient of model with respect to loss function and target

Opto

type Grad (m :: * -> *) a = a -> m (Diff a)

learnGrad :: (MonadSample (a, b) m, Learn a b l, NoState l, LParamMaybe l ~ Just (LParam l), Backprop (LParam l)) => Loss b -> Regularizer (LParam l) -> l -> Grad m (LParam l)

Using a model's deterministic prediction function (with a given loss function), generate a Grad compatible with Numeric.Opto and Numeric.Opto.Run.

learnGradStoch :: (MonadSample (a, b) m, PrimMonad m, Learn a b l, NoState l, LParamMaybe l ~ Just (LParam l), Backprop (LParam l)) => Loss b -> Regularizer (LParam l) -> l -> Gen (PrimState m) -> Grad m (LParam l)

Using a model's stochastic prediction function (with a given loss function), generate a Grad compatible with Numeric.Opto and Numeric.Opto.Run.