Data.Type.Tuple

Description

If you are writing a library that needs to export BVars of tuples, consider using the tuples in this module so that your library can have easy interoperability with other libraries using backprop.

Because of API decisions, backprop and gradBP only work with things with Num instances. However, this disallows default Prelude tuples (without orphan instances from packages like NumInstances).

Until tuples have Num instances in base, this module is intended to be a workaround for situations where:

This comes up often in cases where:

A function wants to return more than one value (BVar s (T2 a b)

You want to uncurry a BVar function to use with backprop and gradBP.

You want to use the useful Prisms automatically generated by the lens library, which use tuples for multiple-constructor fields.

Only 2-tuples and 3-tuples are provided. Any more and you should probably be using your own custom product types, with instances automatically generated from something like one-liner-instances.

Lenses into the fields are provided, but they also work with _1, _2, and _3 from Lens.Micro. However, note that these are incompatible with _1, _2, and _3 from Control.Lens.

You can "construct" a BVar s (T2 a b) with functions like isoVar.

Since: 0.1.1.0

Zero-tuples (unit)

data T0

Unit ('()') with Num, Fractional, and Floating instances.

Be aware that the methods in its numerical instances are all non-strict:

_ + _ = T0
negate _ = T0
fromIntegral _ = T0

Since: 0.1.4.0

Constructors

T0

Two-tuples

data a :# b infixr 1

Strict 2-tuple with Num, Fractional, and Floating instances.

Since: 0.1.1.0

Constructors

!a :# !b infixr 1

pattern (:##) :: (Backprop a, Backprop b, Reifies s W) => BVar s a -> BVar s b -> BVar s (a :# b)

Conversions

t2Tup :: (a :# b) -> (a, b)

Convert to a Haskell tuple.

Forms an isomorphism with tupT2.

tupT2 :: (a, b) -> a :# b

Convert from Haskell tuple.

Forms an isomorphism with t2Tup.

Consumption

uncurryT2 :: (a -> b -> c) -> (a :# b) -> c

Uncurry a function to take in a T2 of its arguments

Since: 0.1.2.0

curryT2 :: ((a :# b) -> c) -> a -> b -> c

Curry a function taking a T2 of its arguments

Since: 0.1.2.0

Lenses

t2_1 :: Lens (a :# b) (a' :# b) a a'

Lens into the first field of a T2. Also exported as _1 from Lens.Micro.

t2_2 :: Lens (a :# b) (a :# b') b b'

Lens into the second field of a T2. Also exported as _2 from Lens.Micro.

N-Tuples

type T = Rec TF

Strict inductive N-tuple with a Num, Fractional, and Floating instances.

It is basically "yet another HList", like the one found in Data.Type.Product and many other locations on the haskell ecosystem. Because it's inductively defined, it has O(n) random indexing, but is efficient for zipping and mapping and other sequential consumption patterns.

It is provided because of its Num instance, making it useful for backproup. Will be obsolete when Product gets numerical instances.

Since: 0.1.5.0

newtype TF a

Constructors

TF

Fields

getTF :: a

pattern (:&&) :: a -> T as -> T (a ': as)

indexT :: Index as a -> T as -> a

Index into a T.

O(i)

Since: 0.1.5.0

Conversions

tOnly :: T '[a] -> a

Extract a singleton T

Forms an isomorphism with onlyT

Since: 0.1.5.0

onlyT :: a -> T '[a]

A singleton T

Forms an isomorphism with tOnly

Since: 0.1.5.0

Lenses

tHead :: Lens (T (a ': as)) (T (b ': as)) a b

Lens into the head of a T

Since: 0.1.5.0

tTail :: Lens (T (a ': as)) (T (a ': bs)) (T as) (T bs)

Lens into the tail of a T

Since: 0.1.5.0

Maybe

type TMaybe = PMaybe TF

data PMaybe (a :: k -> Type) (b :: Maybe k) :: forall k. (k -> Type) -> Maybe k -> Type where

A PMaybe f 'Nothing contains nothing, and a PMaybe f ('Just a) contains an f a.

In practice this can be useful to write polymorphic functions/abstractions that contain an argument that can be "turned off" for different instances.

Constructors

PNothing :: forall k (a :: k -> Type) (b :: Maybe k). PMaybe a (Nothing :: Maybe k)

PJust :: forall k (a :: k -> Type) (b :: Maybe k) (a1 :: k). a a1 -> PMaybe a (Just a1)

Bundled Patterns

pattern TJust :: a -> TMaybe (Just a)

fromPJust :: PMaybe f (Just a) -> f a

type family (a :: Maybe Type) :#? (b :: Maybe Type) :: Maybe Type where ... infixr 1

Equations

Nothing :#? Nothing = Nothing

Nothing :#? (Just b) = Just b

(Just a) :#? Nothing = Just a

(Just a) :#? (Just b) = Just (a :# b)

pattern (:#?) :: (AllConstrainedProd Backprop a, AllConstrainedProd Backprop b, Reifies s W, PureProd Maybe a, PureProd Maybe b) => PMaybe (BVar s) a -> PMaybe (BVar s) b -> PMaybe (BVar s) (a :#? b) infixr 1

splitTupMaybe :: forall f a b. (PureProd Maybe a, PureProd Maybe b) => (forall a' b'. (a ~ Just a', b ~ Just b') => f (a' :# b') -> (f a', f b')) -> PMaybe f (a :#? b) -> (PMaybe f a, PMaybe f b)

tupMaybe :: forall f a b. (forall a' b'. (a ~ Just a', b ~ Just b') => f a' -> f b' -> f (a' :# b')) -> PMaybe f a -> PMaybe f b -> PMaybe f (a :#? b)

pattern MaybB :: (Reifies s W, AllConstrainedProd Backprop a, PureProd Maybe a) => PMaybe (BVar s) a -> BVar s (TMaybe a)