Learn.Neural.Train

Documentation

data Optimizer :: Type -> Type -> Type where

Constructors

MkO :: {..} -> Optimizer a p

Fields

oState :: s

oUpdate :: a -> p -> s -> (p, s)

runOptimizer :: a -> p -> Optimizer a p -> (p, Optimizer a p)

runOptimizer_ :: a -> p -> Optimizer a p -> p

optimizeList :: [a] -> p -> Optimizer a p -> (p, Optimizer a p)

optimizeList_ :: [a] -> p -> Optimizer a p -> p

sgdOptimizer :: forall p as bs c. (Fractional p, Num c, Every Num as, Every Num bs, Known Length as) => Double -> (forall s. OpB s (p ': as) bs) -> LossFunction bs c -> Optimizer (Tuple as, Tuple bs) p

sgdMiniBatchOptimizer :: forall f p as bs c. (Foldable f, Fractional p, Num c, Every Num as, Every Num bs, Known Length as) => Double -> (forall s. OpB s (p ': as) bs) -> LossFunction bs c -> Optimizer (f (Tuple as, Tuple bs)) p

slidingParts :: Nat n -> [a] -> [Vec n a]

slidingPartsSplit :: Nat n -> [(a, b)] -> [(Vec n a, Vec n b)]

slidingPartsLast :: Nat (S n) -> [(a, b)] -> [(Vec (S n) a, b)]

asFeedback :: [a] -> [(a, a)]